### 2021-07-17 #3507 @whedon
[PRE REVIEW]: pyCSEP: A Python Package For Earthquake Forecast Developers
**Submitting author:** @wsavran (<a href="http://orcid.org/0000-0001-5404-0778">William Savran</a>)
**Repository:** <a href="https://github.com/SCECcode/pycsep" target ="_blank">https://github.com/SCECcode/pycsep</a>
**Version:** v0.4.1
