### 2021-07-17 #3507 @danielskatz
👋 @wsavran - your paper doesn't compile- please make the headers match those in the [example](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).  Also, you might be able to fix the possibly missing DOI that whedon suggests, but note that this may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references from branch pycsep_joss` to check again, and the command `@whedon generate pdf from branch pycsep_joss` when the references are right and the .md file is updated to make a new PDF. Whedon commands need to be the first entry in a new comment.
