### 2021-07-18 b7019f2
updated doi for field 2007 in paper.bib
