+++ paper/paper.bib
@@ -4,6 +4,7 @@
    journal = {Seismological Research Letters},
    pages = {1-10},
    url = {http://www.cseptesting.org/documents/files/relm/Field.pdf},
+   DOI = {https://doi.org/10.1785/gssrl.78.1.7},
    year = {2007},
    type = {Journal Article}
 }
