### 2021-07-18 #3507 @danielskatz
You can try to merge https://github.com/SCECcode/pycsep/pull/129 and see if that works - if not, the problem is likely the @ in the email address, which perhaps needs to be protected/quoted in some way.  Perhaps @arfon will know the answer
