### 2021-07-22 717c515
Added percentile argument in plot_args to modify the default confidence interval.
