### 2021-07-22 17b57b7
Added percentile argument in plot_args to modify the default confidence interval. (#131)
