### 2021-07-22 d44b381
Modified comparison plot, to plot simultaneously T and W tests
