### 2021-07-22 b31c6e9
hotfixes for plotting utilities - fixed issue where border was drawn incorrectly - supressed std out message if unable to obtain region_border - axes can be chained for plot_consistency_test_result
