+++ csep/core/regions.py
@@ -744,8 +744,8 @@ class CartesianGrid2D:
             asc.insert(0, poly_min[0])
             asc.insert(0, poly_min[1])
             poly_max = self.polygons[int(row[argmax])].points
-            desc.append(poly_max[3])
             desc.append(poly_max[2])
+            desc.append(poly_max[3])
         # close the loop
         poly = np.array(asc + desc)
         sorted_idx = np.sort(np.unique(poly, return_index=True, axis=0)[1], kind='stable')
