### 2021-07-22 26e6d3a
Made sure, a comparison plot is possible if no results_w  are provided as arg
