### 2021-07-22 b8ce984
Added option to create the figure from previously defined matplotlib.pyplot.ax() object and percentile for W-test confidence interval
