### 2021-07-22 4195021
Modified comparison plot, to plot simultaneously T and W tests (#132)
* Modified comparison plot, to plot simultaneously T and W tests
* Made sure, a comparison plot is possible if no results_w  are provided as arg
* Added option to create the figure from previously defined matplotlib.pyplot.ax() object and percentile for W-test confidence interval
