### 2021-07-22 640ff2e
hotfixes for plotting utilities - grab xticks from axes correly in plot_comparison_test - properly adjust xlim in plot_comparison_test
