### 2021-07-22 5b5843f
Fixed catalog plotting options. The default now for magnitude legends are numpy.linspace(min(mw_range), max(mw_range), 4).   Rounding of the magnitude is to one decimal, but only in the legend text.
