+++ csep/utils/plots.py
@@ -751,12 +751,16 @@ def plot_catalog(catalog, ax=None, show=False, extent=None, set_global=False, pl
     # Legend
     if legend:
         if not mag_ticks:
-            mag_ticks = numpy.round(numpy.linspace(mw_range[0], mw_range[1], 4), 1)
+            mag_ticks = numpy.linspace(mw_range[0], mw_range[1], 4)
+        else:
+            if not numpy.all([ i >= mw_range[0] and i <= mw_range[1] for i in mag_ticks]):
+                print("Magnitude ticks do not lie within the catalog magnitude range")
+
         handles, labels = scatter.legend_elements(prop="sizes",
                                                   num=list(size_map(markersize, mag_ticks, mag_scale)),
                                                   alpha=0.3)
-        ax.legend(handles, mag_ticks,
-                  loc=legend_loc, title=r"Magnitudes",title_fontsize=16,
+        ax.legend(handles, numpy.round(mag_ticks, 1),
+                  loc=legend_loc, title=r"Magnitudes", title_fontsize=16,
                   labelspacing=labelspacing, handletextpad=5, framealpha=False)
 
     if region_border:
