### 2021-07-22 8c42eee
hotfixes for plotting utilities * try to determine which points to plot based on latitude values
