### 2021-07-22 a015af3
hotfixes for plotting utilities * update xlim in plot_comparison_test
