### 2021-07-22 dab5ef9
hotfixes for plotting utilities * update xlim in plot_comparison_test
