### 2021-07-23 528348d
Added option to remove TOP grid labels in spatial maps.  Set 'None' as title name default for plot_spatial_dataset()
