### 2021-07-24 fe858e5
contains hotfixes for plotting utilities (#133)
* hotfixes for plotting utilities
* fixed issue where border was drawn incorrectly
* supressed std out message if unable to obtain region_border
* axes can be chained for plot_consistency_test_result
* grab xticks from axes correly in plot_comparison_test
* properly adjust xlim in plot_comparison_test
* try to determine which points to plot based on latitude values
