### 2021-07-24 e0d0065
Fixed catalog plotting options. (#134)
* Fixed catalog plotting options. The default now for magnitude legends are numpy.linspace(min(mw_range), max(mw_range), 4).   Rounding of the magnitude is to one decimal, but only in the legend text.
* Added option to remove TOP grid labels in spatial maps.  Set 'None' as title name default for plot_spatial_dataset()
