### 2021-07-26 #3507 @kbarnhart
@wsavran thanks for this submission. I have completed my pre-review screening of this submission and will now start soliciting reviewers. If you have reviewer recommendations, please list them here (either list github handle without the @ symbol, or a name if they do not have a github account).

A few minor points regarding the paper to consider.
1. Line 27 - should be "earthquake prediction methods"?
2. Line 53 - should the "the" before "monolithic" be deleted?
3. Line 64 - Europe should be capitalized.
4. Line 84 - provide citations for cartopy and matplotlib.
5. Line 90-91 - provide citations for these packages.
6. Reference section
    - check capitalization (for example, California on line 111, use double curly braces in .bib title entry where necessary).
    - disambiguate Ned Field references (use the identical first name in all .bib entries)

Any questions, clarifications, or concerns... feel free to contact me here or at krbarnhart@usgs.gov
