### 2021-07-26 600e470
refactor polygon to models module
