+++ csep/models.py
@@ -1,6 +1,10 @@
+import matplotlib.path
 import numpy
 
 # CSEP Imports
+import numpy as np
+import pyproj
+
 from csep.utils.time_utils import datetime_to_utc_epoch, epoch_time_to_utc_datetime
 from csep.utils import plots
 
@@ -142,6 +146,7 @@ class CatalogNumberTestResult(EvaluationResult):
         ax = plots.plot_number_test(self, show=show, plot_args=plot_args)
         return ax
 
+
 class CatalogPseudolikelihoodTestResult(EvaluationResult):
 
     def __init__(self, **kwargs):
@@ -158,6 +163,7 @@ class CatalogPseudolikelihoodTestResult(EvaluationResult):
         ax = plots.plot_likelihood_test(self, show=show, plot_args=plot_args)
         return ax
 
+
 class CatalogMagnitudeTestResult(EvaluationResult):
 
     def __init__(self, **kwargs):
@@ -171,6 +177,7 @@ class CatalogMagnitudeTestResult(EvaluationResult):
         plot_args_defaults.update(plot_args)
         ax = plots.plot_magnitude_test(self, show=show, plot_args=plot_args)
         return ax
+
 
 class CatalogSpatialTestResult(EvaluationResult):
 
@@ -190,6 +197,7 @@ class CatalogSpatialTestResult(EvaluationResult):
         ax = plots.plot_spatial_test(self, show=show, plot_args=plot_args)
         return ax
 
+
 class CalibrationTestResult(EvaluationResult):
 
     def __init__(self, **kwargs):
@@ -205,6 +213,7 @@ class CalibrationTestResult(EvaluationResult):
         plot_args_defaults.update(plot_args)
         ax = plots.plot_calibration_test(self, show=show, axes=axes, plot_args=plot_args)
         return ax
+
 
 class EvaluationConfiguration:
     """
@@ -283,3 +292,73 @@ class EvaluationConfiguration:
                 found = True
         if not found:
             self.evaluations.append({'name': name, 'version': version, 'fnames': fnames})
+
+
+class Polygon:
+    """
+    Represents polygons defined through a collection of vertices.
+
+    This polygon is assumed to be 2d, but could contain an arbitrary number of vertices. The path is treated as not being
+    closed.
+    """
+    def __init__(self, points):
+        # instance members
+        self.points = points
+        self.origin = self.points[0]
+
+        # https://matplotlib.org/3.1.1/api/path_api.html
+        self.path = matplotlib.path.Path(self.points)
+
+    def __str__(self):
+        return str(self.origin)
+
+    def contains(self, points):
+        """ Returns a bool array which is True if the path contains the corresponding point.
+
+        Args:
+            points: 2d numpy array
+
+        """
+        nd_points = np.array(points)
+        if nd_points.ndim == 1:
+            nd_points = nd_points.reshape(1,-1)
+        return self.path.contains_points(nd_points)
+
+    def centroid(self):
+        """ return the centroid of the polygon."""
+        c0, c1 = 0, 0
+        k = len(self.points)
+        for p in self.points:
+            c0 = c0 + p[0]
+            c1 = c1 + p[1]
+        return c0 / k, c1 / k
+
+    def get_xcoords(self):
+        return np.array(self.points)[:,0]
+
+    def get_ycoords(self):
+        return np.array(self.points)[:,1]
+
+    @classmethod
+    def from_great_circle_radius(cls, centroid, radius, num_points=10):
+        """
+        Generates a polygon object from a given radius and centroid location.
+
+        Args:
+            centroid: (lon, lat)
+            radius: should be in (meters)
+            num_points: more points is higher resolution polygon
+
+        Returns:
+            polygon
+        """
[…]
