+++ paper.bib
@@ -1,16 +1,16 @@
 @article{Field2007a,
-   author = {Field, Edward},
+   author = {Field, Edward H},
    title = {Overview of the Working Group for the Development of Regional Earthquake Likelihood Models (RELM)},
    journal = {Seismological Research Letters},
    pages = {1-10},
-   url = {http://www.cseptesting.org/documents/files/relm/Field.pdf},
+   url = {http://www.cseptesting.org/documents/files/relm/field.pdf},
    DOI = {https://doi.org/10.1785/gssrl.78.1.7},
    year = {2007},
    type = {Journal Article}
 }
 
 @article{Helmstetter2007a,
-   author = {Helmstetter, Agnes and Kagan, Yan Y and Jackson, David D},
+   author = {Helmstetter, Agnes and Kagan, Y. Y. and Jackson, D. D.},
    title = {High-resolution time-independent grid-based forecast for M$\geq$ 5 earthquakes in California},
    journal = {Seismological Research Letters},
    volume = {78},
@@ -22,7 +22,7 @@
 }
 
 @article{Schorlemmer2007a,
-   author = {Schorlemmer, Danijel and Gerstenberger, MC and Wiemer, S and Jackson, David D and Rhoades, David A},
+   author = {Schorlemmer, Danijel and Gerstenberger, M. C. and Wiemer, S. and Jackson, D. D. and Rhoades, D. A.},
    title = {Earthquake likelihood model testing},
    journal = {Seismological Research Letters},
    volume = {78},
@@ -48,7 +48,7 @@
 }
 
 @article{Zechar2009a,
-   author = {Zechar, Jeremy and Schorlemmer, Danijel and Liukis, Maria and Yu, John and Euchner, Fabian and Maechling, Philip J. and Jordan, Thomas H.},
+   author = {Zechar, Jeremy and Schorlemmer, D. and Liukis, M. and Yu, J. and Euchner, F. and Maechling, P. J. and Jordan, T. H.},
    title = {The Collaboratory for the Study of Earthquake Predictability perspective on computational earthquake science},
    journal = {Concurrency and Computation: Practice and Experience},
    volume = {22},
@@ -61,7 +61,7 @@
 }
 
 @article{Jordan2011a,
-   author = {Jordan, Thomas H and Chen, Y T and Gasparini, Paolo and Madariaga, Raul and Main, Ian and Marzocchi, Warner and Papadopoulos, Gerassimos},
+   author = {Jordan, Thomas H and Chen, Y. T. and Gasparini, P. and Madariaga, R. and Main, I. and Marzocchi, W. and Papadopoulos, G.},
    title = {Operational earthquake forecasting. State of knowledge and guidelines for utilization},
    journal = {Annals of Geophysics},
    DOI = {10.4401/ag-5350},
@@ -101,7 +101,7 @@ https://link.springer.com/content/pdf/10.2478%2Fs11600-011-0013-5.pdf},
 }
 
 @article{Zechar2013a,
-   author = {Zechar, Jeremy and Schorlemmer, Danijel and Werner, Maximilian J and Gerstenberger, M C and Rhoades, David A and Jordan, Thomas H},
+   author = {Zechar, Jeremy and Schorlemmer, D. and Werner, M. J. and Gerstenberger, M. C. and Rhoades, D. A. and Jordan, T. H.},
    title = {Regional Earthquake Likelihood Models I: First-Order Results},
    journal = {Bulletin of the Seismological Society of America},
    volume = {103},
@@ -116,7 +116,7 @@ https://link.springer.com/content/pdf/10.2478%2Fs11600-011-0013-5.pdf},
 
 @article{Cattania2018a,
[…]
