+++ paper.bib
@@ -1,6 +1,6 @@
 @article{Field2007a,
    author = {Field, Edward H},
-   title = {Overview of the Working Group for the Development of Regional Earthquake Likelihood Models (RELM)},
+   title = {Overview of the Working Group for the Development of {{R}}egional {{E}}arthquake {{L}}ikelihood {{M}}odels (RELM)},
    journal = {Seismological Research Letters},
    pages = {1-10},
    url = {http://www.cseptesting.org/documents/files/relm/field.pdf},
@@ -88,7 +88,7 @@ https://link.springer.com/content/pdf/10.2478%2Fs11600-011-0013-5.pdf},
 
 @article{Werner2011a,
    author = {Werner, M. J. and Helmstetter, A. and Jackson, D. D. and Kagan, Y. Y.},
-   title = {High-Resolution Long-Term and Short-Term Earthquake Forecasts for California},
+   title = {High-Resolution Long-Term and Short-Term Earthquake Forecasts for {{C}}alifornia},
    journal = {Bulletin of the Seismological Society of America},
    volume = {101},
    number = {4},
@@ -187,7 +187,7 @@ https://watermark.silverchair.com/srl-2018053.1.pdf?token=AQECAHi208BE49Ooan9kkh
 
 @article{Field2017a,
    author = {Field, Edward H and Jordan, T. H. and Page, M. T. and Milner, K. R. and Shaw, B. E. and Dawson, T. E. and Biasi, G. P. and Parsons, T. and Hardebeck, J. L. and Michael, A. J. and Weldon, R. J. and Powers, P. M. and Johnson, K. M. and Zeng, Y. H. and Felzer, K. R. and van der Elst, N. and Madden, C. and Arrowsmith, R. and Werner, M. J. and Thatcher, W. R.},
-   title = {A Synoptic View of the Third Uniform California Earthquake Rupture Forecast (UCERF3)},
+   title = {A Synoptic View of the third {{U}}niform {{C}}alifornia {{E}}arthquake {{R}}upture {{F}}orecast (UCERF3)},
    journal = {Seismological Research Letters},
    volume = {88},
    number = {5},
