### 2021-07-26 #3507 @wsavran
@kbarnhart, thanks for the feedback on the manuscript. i've asked whedon to compile a new pdf with some changes.

possible reviewers:
* leouieda
* jedbrown
* lheagy

thanks!
