### 2021-08-05 f5d0aa9
Modified and included arguments for spatial plotting functions and filled doc strings with missing variables
- Basemap plotting: Added fontsize arg for grid labels. Added figsize arg. Added manual Zoom Tile level, in case a greater or lower resolution of the map is desired.
- Catalog plotting: Included modified basemap options. Added legend title, framealpha, label fontsize, title fontsize as arguments. Check if given mag_ticks args is list, array or tuple.
- Basemap plotting: Added title_size into correct function args. Added colormap title and ticks fontsize as argument
