+++ csep/utils/plots.py
@@ -553,24 +553,31 @@ def plot_magnitude_histogram(catalogs, comcat, show=True, plot_args=None):
     if show:
         pyplot.show()
 
-def plot_basemap(basemap, extent, ax=None,  coastline=True, borders=False, linecolor='black', linewidth=True,
-                 grid=False, grid_labels=False, set_global=False, show=False, projection=ccrs.PlateCarree(), apprx=False,
-                 central_latitude=0.0):
+def plot_basemap(basemap, extent, ax=None, figsize=None, coastline=True, borders=False, tile_scaling='auto',
+                 set_global=False, projection=ccrs.PlateCarree(), apprx=False, central_latitude=0.0,
+                 linecolor='black', linewidth=True,
+                 grid=False, grid_labels=False, grid_fontsize=None,
+                 show=False):
     """ Wrapper function for multiple cartopy base plots, including access to standard raster webservices
 
      Args:
          basemap (str): Possible values are: stock_img, stamen_terrain, stamen_terrain-background, google-satellite, ESRI_terrain, ESRI_imagery, ESRI_relief, ESRI_topo, ESRI_terrain, or webservice link (see examples in :func:`csep.utils.plots._get_basemap`. Default is None
          extent (list):  [lon_min, lon_max, lat_min, lat_max]
-         show (bool): Flag if the figure is displayed
-         set_global (bool): Display the complete globe as basemap
+         ax (:class:`matplotlib.pyplot.ax`): Previously defined ax object
+         figsize (tuple): If no ax is provided, a tuple of floats can be provided to define figure size
          coastline (str): Flag to plot coastline. default True,
          borders (bool): Flag to plot country borders. default False,
-         linewidth (float): Line width of borders and coast lines. default 1.5,
-         linecolor (str): Color of borders and coast lines. default 'black',
-         grid (bool): Draws a grid in the basemap
-         grid_labels (bool): Annotate grid values
+         tile_scaling (str/int): Zoom level (1-12) of the basemap tiles. If 'auto', is automatically derived from extent
+         set_global (bool): Display the complete globe as basemap
+         projection (:class:`cartopy.crs.Projection`): Projection to be used in the basemap
          apprx (bool): If true, approximates transformation by setting aspect ratio of axes based on middle latitude
          central_latitude (float): average latitude from plotting region
+         linecolor (str): Color of borders and coast lines. default 'black',
+         linewidth (float): Line width of borders and coast lines. default 1.5,
+         grid (bool): Draws a grid in the basemap
+         grid_labels (bool): Annotate grid values
+         grid_fontsize (float): Font size of the grid x and y labels
+         show (bool): Flag if the figure is displayed
 
      Returns:
          :class:`matplotlib.pyplot.ax` object
@@ -579,14 +586,14 @@ def plot_basemap(basemap, extent, ax=None,  coastline=True, borders=False, linec
     if ax is None:
         if apprx:
             projection = ccrs.PlateCarree()
-            fig = pyplot.figure()
+            fig = pyplot.figure(figsize=figsize)
             ax = fig.add_subplot(111, projection=projection)
             # Set plot aspect according to local longitude-latitude ratio in metric units
             # (only compatible with plain PlateCarree "projection")
             LATKM = 110.574  # length of a ° of latitude [km]; constant --> ignores Earth's flattening
[…]
