### 2021-08-10 #3507 @kbarnhart
👋 @calum-chamberlain, @megies, @barsch, @mhearne-usgs would you be interested and able to provide a review of a manuscript at the Journal of Open Source Software (JOSS). The submission I'd like you to review is titled: " pyCSEP: A Python Package For Earthquake Forecast Developers" by William H. Savran, Maximilian J. Werner, Danijel Schorlemmer, and Philip J. Maechling and the submission repository is at: https://github.com/SCECcode/pycsep. Given your extensive experience in earthquake seismology and python development, I would value your review of this submission.

A bit of background about JOSS and its review process. JOSS is a free, open-source, community driven and developer-friendly online journal (no publisher is seeking to raise revenue from the volunteer labor of researchers!). The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html. Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review).

JOSS has worked hard to recognize that because of the challenges of the COVID-19 pandemic, all members of the research community may have reduced availability. At present we are asking reviewers to complete their reviews within 6 weeks.

If you are interested and able, please review the [JOSS Conflict of Interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). Otherwise I’d value any recommendations you may have for other reviewers.
