### 2021-08-14 9d0cb07
added get_valid_midpoints to gridded forecast
