### 2021-08-16 #3507 @cjhopp
Hi @kbarnhart , apologies for the delay in responding, and thanks for reaching out! In the future, I'd be happy to review for JOSS, but I have to decline this time around. Many weeks of impending fieldwork.
