+++ CITATION.cff
@@ -1,0 +1,8 @@
+cff-version: 1.1.0
+message: "If you use this software, please cite it as below."
+authors:
+  - family-names: Savran
+    given-names: William
+title: "pyCSEP"
+version: 0.4.1
+date-released: 2021-04-20 
