### 2021-08-16 #3507 @baagaard-usgs
Katy,

CSEP is a bit outside my area of expertise. Morgan Page, Nicholas van der Elst, and Andrew Michael are in a much better position than me to review the scientific content and probably the software content as well.

Brad
