### 2021-08-17 #3507 @kbarnhart
Thanks to @brendanjmeade @baagaard-usgs and @MatteoTaroniINGV for your quick responses and the recommendations.
