### 2021-08-22 #3507 @kbarnhart
👋 @angystallone a friendly reminder to consider serving as a reviewer for "pyCSEP: A Python Package For Earthquake Forecast Developers" by William H. Savran, Maximilian J. Werner, Danijel Schorlemmer, and Philip J. Maechling at the Journal of Open Source Software. I'll also send you a message via email. Many thanks for considering.
