### 2021-08-24 540d8fd
add serialization to regions for json formatted catalog - supports new region versions e.g., QuadTreeRegion - added region comparison using to_dict method in regions
