### 2021-08-24 065ae2d
refactor polygon to models module (#135)
