### 2021-08-24 dac5f0f
fixed issue where tests fail if catalog does not have an associated region
