### 2021-08-24 7686eda
update README.md (#139)
fixes #106
