### 2021-08-24 e03514f
added flag to offset grids from files by dh / 2 output computing metrics by default when iterating over catalog forecasts cleaned up docstrings
