+++ csep/core/regions.py
@@ -14,13 +14,21 @@ from csep.utils.scaling_relationships import WellsAndCoppersmith
 from csep.models import Polygon
 
 
-def california_relm_collection_region(dh_scale=1, magnitudes=None, name="relm-california-collection"):
+def california_relm_collection_region(dh_scale=1, magnitudes=None, name="relm-california-collection", use_midpoint=True):
     """ Return collection region for California RELM testing region
 
-        Args:
-            dh_scale (int): factor of two multiple to change the grid size
-            mangitudes (array-like): array representing the lower bin edges of the magnitude bins
-            name (str): human readable identifer
+    Args:
+        dh_scale (int): factor of two multiple to change the grid size
+        mangitudes (array-like): array representing the lower bin edges of the magnitude bins
+        name (str): human readable identifer
+        use_midpoints (bool): if true, treat values in file as midpoints. default = true.
+
+    Returns:
+        :class:`csep.core.spatial.CartesianGrid2D`
+
+    Raises:
+        ValueError: dh_scale must be a factor of two
+
     """
     if dh_scale % 2 != 0 and dh_scale != 1:
         raise ValueError("dh_scale must be a factor of two or dh_scale must equal unity.")
@@ -29,8 +37,11 @@ def california_relm_collection_region(dh_scale=1, magnitudes=None, name="relm-ca
     dh = 0.1
     root_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
     filepath = os.path.join(root_dir, 'artifacts', 'Regions', 'RELMCollectionArea.dat')
-    midpoints = numpy.loadtxt(filepath)
-    origins = midpoints - dh / 2
+    points = numpy.loadtxt(filepath)
+    if use_midpoint:
+        origins = numpy.array(points) - dh / 2
+    else:
+        origins = numpy.array(points)
 
     if dh_scale > 1:
         origins = increase_grid_resolution(origins, dh, dh_scale)
@@ -45,7 +56,7 @@ def california_relm_collection_region(dh_scale=1, magnitudes=None, name="relm-ca
 
     return relm_region
 
-def california_relm_region(dh_scale=1, magnitudes=None, name="relm-california"):
+def california_relm_region(dh_scale=1, magnitudes=None, name="relm-california", use_midpoint=True):
     """
     Returns class representing California testing region.
 
@@ -71,8 +82,11 @@ def california_relm_region(dh_scale=1, magnitudes=None, name="relm-california"):
     root_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
     filepath = os.path.join(root_dir, 'artifacts', 'Regions', 'csep-forecast-template-M5.xml')
     csep_template = os.path.expanduser(filepath)
-    midpoints, dh = parse_csep_template(csep_template)
-    origins = numpy.array(midpoints) - dh / 2
+    points, dh = parse_csep_template(csep_template)
+    if use_midpoint:
+        origins = numpy.array(points) - dh / 2
+    else:
+        origins = numpy.array(points)
 
     if dh_scale > 1:
         origins = increase_grid_resolution(origins, dh, dh_scale)
@@ -87,7 +101,7 @@ def california_relm_region(dh_scale=1, magnitudes=None, name="relm-california"):
 
     return relm_region
 
-def italy_csep_region(dh_scale=1, magnitudes=None, name="csep-italy"):
+def italy_csep_region(dh_scale=1, magnitudes=None, name="csep-italy", use_midpoint=True):
     """
         Returns class representing Italian testing region.
 
@@ -98,6 +112,8 @@ def italy_csep_region(dh_scale=1, magnitudes=None, name="csep-italy"):
             dh_scale: can resample this grid by factors of 2
[…]
