### 2021-08-25 #3658 @whedon
[REVIEW]: pyCSEP: A Python Package For Earthquake Forecast Developers
**Submitting author:** @wsavran (<a href="http://orcid.org/0000-0001-5404-0778">William Savran</a>)
**Repository:** <a href="https://github.com/SCECcode/pycsep" target ="_blank">https://github.com/SCECcode/pycsep</a>
**Version:** v0.5.2

@nvanderelst and @mbarall, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
