### 2021-08-25 #3658 @kbarnhart
@nvanderelst - thanks for agreeing to review this submission to JOSS. In the prior comments there is a checklist that you can use to guide you through your review. There are also instructions to manage your github notifications so you don't get all notifications for all JOSS reviews.

We have an automatic reminder set up in two weeks to ask you how the review is going. At present we request that reviewers complete their reviews within 6 weeks. I know you indicated you would be able to get to this sometime in late September (totally fine). JOSS is trying to be mindful of changes people have experienced due to COVID-19.

As you work through your review, if there are any issues that come up, please make an issue in the PyCEP repository, and link to this issue (paste `openjournals/joss-reviews/issues/3658` into the issue you create). That way most of the discussion can occur on in-repo issues.

If you have any questions, please let me know (tag me here or email krbarnhart@usgs.gov).

I plan to find a second reviewer for this submission and will manually edit this page to provide them a checklist once they have been identified.
