### 2021-08-30 #3658 @kbarnhart
@mbarall - thanks for agreeing to review this submission to JOSS. In the prior comments there is a checklist that you can use to guide you through your review. There are also instructions to manage your github notifications so you don't get all notifications for all JOSS reviews.

We have an automatic reminder set up in two weeks to ask you how the review is going. At present we request that reviewers complete their reviews within 6 weeks. You indicated to me that you expect your review to take a few weeks longer, which is totally fine. JOSS is trying to be mindful of changes people have experienced due to COVID-19. Please update me if anything else comes up.

As you work through your review, if there are any issues that come up, please make an issue in the PyCEP repository, and link to this issue (paste openjournals/joss-reviews/issues/3658 into the issue you create). That way most of the discussion can occur on in-repo issues.

Thanks again for contributing a review to JOSS.
