### 2021-09-22 #3658 @kbarnhart
@wsavran @nvanderelst @mbarall I wanted to let you know I'll be in the field with very limited internet access from Sept 24 through Sept 30. If any questions come up during that time, I'll respond when I return.

@nvanderelst, @mbarall I know that we are getting close (or are in) the time you expected to be able to do this review. Please let me know if you have any questions as you undertake your reviews. Thanks again for being willing to serve as a reviewer for JOSS.
