### 2021-10-05 f42a502
added support for forecasts with single magnitude bin fixes #140
