### 2021-10-06 dc48a52
fixes issue #140 to add support for spatial forecasts (#142)
* fixed imports for spatial_likelihood_function
* added support for forecasts with single magnitude bin
fixes #140
* added additional tests
