### 2021-10-26 e02e7da
remove numpy pinned version from docs workflow
