+++ .github/workflows/build-sphinx.yml
@@ -22,7 +22,7 @@ jobs:
           sudo apt-get install libgeos-dev
           sudo apt-get install python3-sphinx
           python -m pip install --upgrade pip
-          pip install numpy==1.18.4
+          pip install numpy
           pip install wheel
           pip install pytest pytest-cov
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
