### 2021-10-27 82a5400
added ci workflow to build documentation on pull request added option to increase border pading around catalog plots added function to label figures for publication changed test_global_binning to use california_relm_region to preserve memory
