### 2021-10-27 8ac84dc
fixed bug in workflow yaml file added checklist for pycsep pull requests
