### 2021-10-27 78ec12c
bug in docs run script. removed manual triggering.
