+++ .github/workflows/build-sphinx-no-deploy.yml
@@ -1,9 +1,6 @@
 name: docs
 
 on:
-  workflow_dispatch:
-    tags:
-      descriptions: 'Manually trigger documentation build'
   pull_request:
     branches: [ master ]
 
