### 2021-10-27 6c485da
testing docs build with conda instead of pip
