### 2021-10-27 f981719
testing docs build with conda instead of pip
