+++ .github/workflows/build-sphinx-no-deploy.yml
@@ -14,9 +14,8 @@ jobs:
       - name: Install dependencies
         run: |
           $CONDA/bin/conda env update --file requirements.yml --name base
-      - name: Install PyCSEP
-        run: |
-          pip install -e .
+          $CONDA/bin/conda install conda-build
+          $CONDA/bin/conda develop .
       - name: Build documentation
         run: |
           make -C docs clean
