+++ .github/workflows/build-sphinx-no-deploy.yml
@@ -14,10 +14,11 @@ jobs:
       - name: Install dependencies
         run: |
           $CONDA/bin/conda env update --file requirements.yml --name base
+          $CONDA/bin/conda activate base
           $CONDA/bin/pip install -e .
+        shell: bash
       - name: Build documentation
         run: |
           make -C docs clean
           make -C docs html
         shell: bash
-
