+++ .github/workflows/build-sphinx-no-deploy.yml
@@ -5,6 +5,7 @@ on:
 jobs:
   build_docs:
     runs-on: ubuntu-latest
+    shell: bash
     steps:
       - uses: actions/checkout@v2
       - name: Set up Python 3.7
@@ -19,4 +20,3 @@ jobs:
         uses: ammaraskar/sphinx-action@master
         with:
           docs-folder: "docs/"
-        shell: bash
