+++ .github/workflows/build-sphinx-no-deploy.yml
@@ -7,6 +7,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
+      - name: Install pyCSEP
       - uses: conda-incubator/setup-miniconda@v2
         with:
           activate-environment: csep-dev
