+++ .github/workflows/build-sphinx-no-deploy.yml
@@ -7,18 +7,13 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-      - name: Install pyCSEP
       - uses: conda-incubator/setup-miniconda@v2
         with:
           activate-environment: csep-dev
           python-version: 3.9
           channels: conda-forge
           environment-file: requirements.yml
+          auto-active-base: false
       - run: |
           conda install conda-build
           conda develop .
-      - name: Build documentation
-      - uses: ammaraskar/sphinx-action@master
-        with:
-          docs-folder: "docs/"
-
