+++ .github/workflows/build-sphinx-no-deploy.yml
@@ -7,16 +7,26 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-      - uses: conda-incubator/setup-miniconda@v2
+      - name: Set up Python 3.7
+        uses: actions/setup-python@v2
         with:
-          activate-environment: csep-dev
-          python-version: 3.9
-          channels: conda-forge
-          environment-file: requirements.yml
-          auto-active-base: false
-      - run: |
-          conda install conda-build
-          conda develop .
-      - uses: ammaraskar/sphinx-action@master
-        with:
-          docs-folder: "docs/"
+          python-version: 3.7
+      - name: Install dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get install libproj-dev proj-data proj-bin
+          sudo apt-get install libgeos-dev
+          sudo apt-get install python3-sphinx
+          python -m pip install --upgrade pip
+          pip install numpy
+          pip install wheel
+          pip install pytest pytest-cov
+          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+          pip install pillow
+      - name: Install PyCSEP
+        run: |
+          python setup.py install
+      - name: Build documentation
+        run: |
+          make -C docs clean
+          make -C docs html
