### 2021-10-27 74515bd
updated build and deploy docs script
