+++ CONTRIBUTING.md
@@ -7,9 +7,9 @@ answer some questions.
 ## Getting Started
 
 * Make sure you have an active GitHub account
-* Download and install git
-* Read the git documentation
-* Install a **development version** of PyCSEP
+* Download and install `git`
+* Read git documentaion if you aren't familiar with `git`
+* Install the **development version** of PyCSEP
 * If you haven't worked with git Forks before, make sure to read the documentation linked here:
 [some helping info](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/working-with-forks).
 
@@ -24,20 +24,21 @@ We recommend using `conda` to install the development environment.
     conda env create -f requirements.yml
     conda activate csep-dev
     pip install -e .[all]
-    # add upstream repository with default repository
+    # add upstream repository
     git remote add upstream https://github.com/SCECCode/pycsep.git
     
-Note: use the commend `conda deactivate` to go back to your regular environment.
+Note: use the command `conda deactivate` to go back to your regular environment when you are done working with pyCSEP.
 
 ## Submitting a Pull Request
 
 ### Some notes for starting a pull request
 
-Pull requests are great! Please submit them to us! Here's how:
+Pull requests are how we use your changes to the code! Please submit them to us! Here's how:
 
 1. Make a new branch. For features/additions base your new branch at `master`.
-2. Add a test! Only pull requests for documentation, refactoring, or plotting features do not require a test.
+2. Make sure to add tests! Only pull requests for documentation, refactoring, or plotting features do not require a test.
 3. Also, documentation must accompany new feature requests.
+    - Note: We would really appreciate pull requests that help us improve documentation.
 3. Make sure the tests pass. Run `./run_tests.sh` in the top-level directory of the repo.
 4. Push your changes to your fork and submit a pull request. Make sure to set the branch to `pycsep:master`.
 5. Wait for our review. There may be some suggested changes or improvements. Changes can be made after
@@ -49,11 +50,11 @@ This allows us to provide feedback during development and to answer any question
 Also, if you find pyCSEP to be useful, but don't want to contribute to the code we highly encourage updates to the documentation!
 
 Please make sure to set the correct branch for your pull request. Also, please do not include large files in your pull request.
-If you feel that you need to add large files, let us know and we can figure something out.
+If you feel that you need to add large files, such as a benchmark forecast, let us know and we can figure something out.
 
 ### Tips to get your pull request accepted quickly
 
-1. Any new feature that contains calculations must contain a unit-test to ensure that the calculations are doing what you
+1. Any new feature that contains calculations must contain unit-tests to ensure that the calculations are doing what you
 expect. Some exceptions to this are documentation changes and new plotting features. 
 2. Documentation should accompany any new feature additions into the package.
     * Plotting functions should provide a sphinx-gallery example, which can be found [here](https://github.com/SCECcode/pycsep/blob/master/examples/tutorials/catalog_filtering.py).
@@ -62,7 +63,10 @@ expect. Some exceptions to this are documentation changes and new plotting featu
 3. pyCSEP uses pytest as a test runner. Add new tests to the `tests` folder in an existing file or new file starting matching `test_*.py`
 4. New scientific capabilities that are not previously published should be presented to the CSEP science group as part of a 
 science review. This will consist of a presentation that provides a scientific justification for the feature.
-
+5. Code should follow the [pep8](https://pep8.org/) style-guide.
[…]
