### 2021-10-28 aa078b3
ci tools to build documentation in pr and test fix for oom issues in build (#145)
* added ci workflow to build documentation on pull request
* added option to increase border pading around catalog plots
* added function to label figures for publication
* changed test_global_binning to use california_relm_region to preserve memory
* fixed bug in workflow yaml file
* added checklist for pycsep pull requests
* bug in docs run script. removed manual triggering.
* updates to CONTRIBUTING.md
