### 2021-10-31 72746fc
added verbose option to catalog based tests added get_event_counts method to CatalogForecast added plotting arguments for more granular customization
