+++ README.md
@@ -1,13 +1,12 @@
 # PyCSEP: Collaboratory for the Study of Earthquake Predictability
-
 ![](http://hypocenter.usc.edu/research/badges/CSEP2_Logo_CMYK.png)
-
 ![Python version](http://hypocenter.usc.edu/research/badges/pycsep-python.svg)
 ![Python application](https://github.com/SCECCode/csep2/workflows/Python%20application/badge.svg)
 [![Build sphinx documentation](https://github.com/SCECCode/csep2/workflows/Build%20sphinx%20documentation/badge.svg)](https://cseptesting.org)
 [![codecov](https://codecov.io/gh/SCECcode/pycsep/branch/master/graph/badge.svg?token=HTMKM29MAU)](https://codecov.io/gh/SCECcode/pycsep)
 
-The PyCSEP tools help earthquake forecast model developers evaluate their forecasts with the goal of understanding
+# Description:
+The PyCSEP Toolkit helps earthquake forecast model developers evaluate their forecasts with the goal of understanding
 earthquake predictability.
 
 PyCSEP should:
@@ -17,92 +16,66 @@ PyCSEP should:
 4. Promote open-science ideas by ensuring transparency and availability of scientific code and results.
 5. Curate benchmark models and data sets for modelers to conduct retrospective experiments of their forecasts.
 
-## Using Conda
+# Table of Contents:
+1. [Software Documentation](https://docs.cseptesting.org)
+2. [Installation](#installation)
+3. [Usage](#usage)
+4. [Contributing](#contributing)
+5. [Change Log](https://github.com/SCECcode/pycsep/blob/master/CHANGELOG.txt)
+6. [Credits](#credit)
+7. [License](#license)
 
-The easiest way to install PyCSEP is using `conda`. It can also be installed using `pip` or built from source. 
-If you plan on contributing to this package, visit the 
-[contribution guidelines](https://github.com/SCECcode/pycsep/blob/master/CONTRIBUTING.md) for 
-installation instructions. We recommend creating a `conda` environment for this installation, but this is not a required step.
+# Installation:
+* [Detailed PyCSEP Installation Instructions](https://docs.cseptesting.org/getting_started/installing.html)
 
-    conda create -n pycsep
-    conda activate pycsep
-    
-Install `pycsep`
+PyCSEP can be installed multiple ways in multiple computing environments. PyCSEP can be installed using conda or pip package managers or
+from source code found in the PyCSEP github repo. Researchers interested in contributing should install pyCSEP from source code. 
+PyCSEP depends on the following software packages. 
+These which may be installed automatically, or manually, depending on the installation method used.
+* Python 3.7 or later (https://python.org)
+* NumPy 1.10 or later (https://numpy.org)  
+* GEOS 3.3.3 or later (https://trac.osgeo.org/geos/)  
+* PROJ 4.9.0 or later (https://proj4.org/)  
 
-    conda install --channel conda-forge pycsep
-    
-## Using Pip
+# Usage: 
+Once installed, PyCSEP methods can be invoked from python code by importing package csep. PyCSEP provides objects and utilities related to several key concepts:
+* Earthquake Catalogs
+* Earthquake Forecasts
+* Earthquake Forecast Evaluations
+* Regions
 
-Before this installation will work, you must **first** install the following system dependencies. The remaining dependencies
-should be installed by the installation script. To help manage dependency issues, we recommend using virtual environments 
-like `virtualenv`.
+An example of PyCSEP catalog operations is shown below:
+<pre>
+import csep
+from csep.core import regions
+from csep.utils import time_utils, comcat
+start_time = csep.utils.time_utils.strptime_to_utc_datetime('2019-01-01 00:00:00.0')
[…]
