+++ CREDITS.md
@@ -1,0 +1,15 @@
+PyCSEP development is a group effort.
+
+Contributors:
+* William Savran, Southern California Earthquake Center
+* Pablo Iturrieta
+* khawajasim 
+* gen2
+* mherrmann3
+* bayonato89 
+* levin422
+* Max Werner, Southern California Earthquake Center
+* Danijel Schorlemmner, 
+* Philip Maechling, Southern California Earthquake Center
+
+Thanks to everyone for all your contributions! 
