+++ LICENSE.txt
@@ -1,4 +1,4 @@
-Copyright (c) 2020, The University of Southern California
+Copyright (c) 2021, The University of Southern California
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
