+++ README.md
@@ -66,16 +66,8 @@ Software support for PyCSEP is provided by that Southern California Earthquake C
 We welcome contributions to the PyCSEP Toolkit.  If you would like to contribute to this package, including software, tests, and documentation, please visit the [contribution guidelines](https://github.com/SCECcode/pycsep/blob/master/CONTRIBUTING.md) for guidelines on how pyCSEP commuinty development instructions.
 
 # Credits:
-The following developers have contributed to the development of the PyCSEP Toolkit. 
-* William Savran, Southern California Earthquake Center
-* Pablo Iturrieta
-* khawajasim 
-* gen2
-* mherrmann3
-* bayonato89 
-* levin422
-* Max Werner, Southern California Earthquake Center
-* Philip Maechling, Southern California Earthquake Center
+Development of PyhCSEP is a group effort. A list of developers that have contributed to the PyCSEP Toolkit 
+are listed in the [Credits](CREDITS.md) file in this repository.
 
 # License:
 The pyCSEP software is distributed under the BSD 3-Clause open-source license. Please see the [LICENSE.txt file](LICENSE.txt) for more information.
