### 2021-11-01 6b825f7
Updated wording about installations
