### 2021-11-01 c5a851f
standardized name as PyCSEP in this file
