### 2021-11-01 5a5ca2e
Added link to Code of Conduct guidelines
