+++ README.md
@@ -63,7 +63,9 @@ Software support for PyCSEP is provided by that Southern California Earthquake C
 2. Email Contact: software@scec.usc.edu
 
 # Contributing:
-We welcome contributions to the PyCSEP Toolkit.  If you would like to contribute to this package, including software, tests, and documentation, please visit the [contribution guidelines](https://github.com/SCECcode/pycsep/blob/master/CONTRIBUTING.md) for guidelines on how PyCSEP commuinty development instructions.
+We welcome contributions to the PyCSEP Toolkit.  If you would like to contribute to this package, including software, tests, and documentation, 
+please visit the [contribution guidelines](https://github.com/SCECcode/pycsep/blob/master/CONTRIBUTING.md) for guidelines on how to contribute to PyCSEP development.
+PyCSEP contributors agree to abide by the code of conduct found in our [Code of Conduct](CODE_OF_CONDUCT.md) guidelines.
 
 # Credits:
 Development of PyCSEP is a group effort. A list of developers that have contributed to the PyCSEP Toolkit 
