### 2021-11-01 7844be3
updated install instructions
