### 2021-11-01 a85ed40
added codemeta.json file and updated CREDITS.md
