### 2021-11-01 3939290
updated list of contributors in CREDITS.md
