+++ CREDITS.md
@@ -2,13 +2,12 @@ PyCSEP development is a group effort.
 
 Contributors:
 * William Savran, Southern California Earthquake Center
-* Pablo Iturrieta
-* Muhammad Asim Khawaja
-* Thomas Beutin
-* Toño Bayona, University of Bristol
-* gen2
-* mherrmann3
-* levin422
+* Pablo Iturrieta, GFZ Potsdam
+* Khawaja Asim, GFZ Potsdam
+* Thomas Beutin, GFZ Potsdam
+* Jose Bayona, University of Bristol
+* Marcus Hermann, University of Naples 'Frederico II'
+* Edric Pauk, Southern California Earthquake Center
 * Max Werner, University of Bristol
 * Danijel Schorlemmner, GFZ Potsdam
 * Philip Maechling, Southern California Earthquake Center
