### 2021-11-01 51690bc
Changed np for numpy in the poisson_evaluations.py file
