+++ csep/utils/plots.py
@@ -1,0 +1,1627 @@
+import time
+
+# Third-party imports
+import numpy
+import string
+
+import scipy.stats
+import matplotlib
+from matplotlib import cm
+from matplotlib.collections import PatchCollection
+import matplotlib.pyplot as pyplot
+import cartopy
+import cartopy.crs as ccrs
+from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
+from cartopy.io import img_tiles
+# PyCSEP imports
+from csep.utils.constants import SECONDS_PER_DAY, CSEP_MW_BINS
+from csep.utils.calc import bin1d_vec
+from csep.utils.time_utils import datetime_to_utc_epoch
+
+
+"""
+This module contains plotting routines that generate figures for the stochastic event sets produced from
+CSEP2 experiments.
+
+Right now functions dont have consistent signatures. That means that some functions might have more functionality than others
+while the routines are being developed.
+
+TODO: Add annotations for other two plots.
+TODO: Add ability to plot annotations from multiple catalogs. Esp, for plot_histogram()
+IDEA: Same concept mentioned in evaluations might apply here. The plots could be a common class that might provide
+      more control to the end user.
+IDEA: Since plotting functions are usable by these classes only that don't implement iter routines, maybe make them a class
+      method. like data.plot_thing()
+"""
+
+def plot_cumulative_events_versus_time_dev(xdata, ydata, obs_data, plot_args, show=False):
+    """
+
+
+    Args:
+        xdata (ndarray): time bins for plotting shape (N,)
+        ydata (ndarray or list like): ydata for plotting; shape (N,5) in order 2.5%Per, 25%Per, 50%Per, 75%Per, 97.5%Per
+        obs_data (ndarry): same shape as xdata
+        plot_args:
+        show:
+
+    Returns:
+
+    """
+    figsize = plot_args.get('figsize', None)
+    sim_label = plot_args.get('sim_label', 'Simulated')
+    obs_label = plot_args.get('obs_label', 'Observation')
+    legend_loc = plot_args.get('legend_loc', 'best')
+    title = plot_args.get('title', 'Cumulative Event Counts')
+    xlabel = plot_args.get('xlabel', 'Days')
+
+    fig, ax = pyplot.subplots(figsize=figsize)
+    try:
+        fifth_per = ydata[0,:]
+        first_quar = ydata[1,:]
+        med_counts = ydata[2,:]
+        second_quar = ydata[3,:]
+        nine_fifth = ydata[4,:]
+    except:
+        raise TypeError("ydata must be a [N,5] ndarray.")
+    # plotting
+
+    ax.plot(xdata, obs_data, color='black', label=obs_label)
+    ax.plot(xdata, med_counts, color='red', label=sim_label)
+    ax.fill_between(xdata, fifth_per, nine_fifth, color='red', alpha=0.2, label='5%-95%')
+    ax.fill_between(xdata, first_quar, second_quar, color='red', alpha=0.5, label='25%-75%')
+    ax.legend(loc=legend_loc)
+    ax.set_xlabel(xlabel)
+    ax.set_ylabel('Cumulative event count')
+    ax.set_title(title)
+    # pyplot.subplots_adjust(right=0.75)
+    # annotate the plot with information from data
+    # ax.annotate(str(observation), xycoords='axes fraction', xy=xycoords, fontsize=10, annotation_clip=False)
+    # save figure
+    filename = plot_args.get('filename', None)
+    if filename is not None:
+        fig.savefig(filename + '.pdf')
+        fig.savefig(filename + '.png', dpi=300)
+    # optionally show figure
+    if show:
+        pyplot.show()
+
[…]
