### 2021-11-02 b496911
Modify plots.py on top of the lastes master version
