### 2021-11-02 57b8503
Adding documentations of the two new functions
