### 2021-11-03 aa65b54
small modifications to README.md * updated links to images * changed example snippet to include plotting * minor text changes
