+++ README.md
@@ -1,18 +1,18 @@
-# PyCSEP: Collaboratory for the Study of Earthquake Predictability
-![](https://g-c662a6.a78b8.36fe.data.globus.org/csep/badges/CSEP2_Logo_CMYK.png)
-![Python version](https://g-c662a6.a78b8.36fe.data.globus.org/csep/badges/pycsep-python.svg)
+# pyCSEP: Collaboratory for the Study of Earthquake Predictability
+![](https://i.postimg.cc/Bb60rVQP/CSEP2-Logo-CMYK.png)
+![Python version](https://gist.githubusercontent.com/wsavran/efce311162c32460336a4f9892218532/raw/1b9c060efd1c6e52eb53f82d4249107417d6a5ec/pycsep_python_badge.svg)
 ![Python application](https://github.com/SCECCode/csep2/workflows/Python%20application/badge.svg)
 [![Build sphinx documentation](https://github.com/SCECCode/csep2/workflows/Build%20sphinx%20documentation/badge.svg)](https://cseptesting.org)
 [![codecov](https://codecov.io/gh/SCECcode/pycsep/branch/master/graph/badge.svg?token=HTMKM29MAU)](https://codecov.io/gh/SCECcode/pycsep)
 
 # Description:
-The PyCSEP Toolkit helps earthquake forecast model developers evaluate their forecasts with the goal of understanding
+The pyCSEP Toolkit helps earthquake forecast model developers evaluate their forecasts with the goal of understanding
 earthquake predictability.
 
-PyCSEP should:
+pyCSEP should:
 1. Help modelers become familiar with formats, procedures, and evaluations used in CSEP Testing Centers.
 2. Provide vetted software for model developers to use in their research.
-3. Provide quantative and visual tools to assess earthquake forecast quality.
+3. Provide quantitative and visual tools to assess earthquake forecast quality.
 4. Promote open-science ideas by ensuring transparency and availability of scientific code and results.
 5. Curate benchmark models and data sets for modelers to conduct retrospective experiments of their forecasts.
 
@@ -22,54 +22,64 @@ PyCSEP should:
 3. [Usage](#usage)
 4. [Contributing](#contributing)
 5. [Change Log](https://github.com/SCECcode/pycsep/blob/master/CHANGELOG.txt)
-6. [Credits](#credit)
+6. [Credits](#credits)
 7. [License](#license)
 
 # Installation:
-PyCSEP can be installed in several ways. It can be installed using conda or pip package managers or from the 
-source code found in the PyCSEP github repo. Researchers interested in contributing to PyCSEP development should 
-install PyCSEP from source code. PyCSEP depends on the following software packages. 
+pyCSEP can be installed in several ways. It can be installed using conda or pip package managers or from the 
+source code found in the pyCSEP github repo. Researchers interested in contributing to pyCSEP development should 
+install pyCSEP from source code. pyCSEP depends on the following software packages. 
 These which may be installed automatically, or manually, depending on the installation method used.
 * Python 3.7 or later (https://python.org)
-* NumPy 1.10 or later (https://numpy.org)  
-* GEOS 3.3.3 or later (https://trac.osgeo.org/geos/)  
-* PROJ 4.9.0 or later (https://proj4.org/)  
+* NumPy 1.21.3 or later (https://numpy.org)
+* SciPy 1.7.1 or later (https://scipy.org)
+* pandas 1.3.4 or later (https://pandas.pydata.org)
+* cartopy 0.20.0 or later (https://scitools.org.uk/cartopy/docs/latest)
+* GEOS 3.7.2 or later (https://trac.osgeo.org/geos/)
+* PROJ 8.0.0 or later (https://proj.org/)
 
[…]
