### 2021-11-03 9e53c0a
Updated readme (#147)
* added extension to file and updatd to standard format readme.md
* Moved credits to Credits file
* Updated wording about installations
* standardized name as PyCSEP in this file
* Added link to Code of Conduct guidelines
* updated url to carc public filesystem
* updated install instructions
* added codemeta.json file and updated CREDITS.md
* updated list of contributors in CREDITS.md
* small modifications to README.md
* updated links to images
* changed example snippet to include plotting
* minor text changes

Co-authored-by: William Savran <wsavran@usc.edu>
