+++ CREDITS.md
@@ -1,0 +1,15 @@
+PyCSEP development is a group effort.
+
+Contributors:
+* William Savran, Southern California Earthquake Center
+* Pablo Iturrieta, GFZ Potsdam
+* Khawaja Asim, GFZ Potsdam
+* Thomas Beutin, GFZ Potsdam
+* Jose Bayona, University of Bristol
+* Marcus Hermann, University of Naples 'Frederico II'
+* Edric Pauk, Southern California Earthquake Center
+* Max Werner, University of Bristol
+* Danijel Schorlemmner, GFZ Potsdam
+* Philip Maechling, Southern California Earthquake Center
+
+Thanks to everyone for all your contributions! 
