+++ README.md
@@ -1,108 +1,85 @@
-# PyCSEP: Collaboratory for the Study of Earthquake Predictability
-
-![](http://hypocenter.usc.edu/research/badges/CSEP2_Logo_CMYK.png)
-
-![Python version](http://hypocenter.usc.edu/research/badges/pycsep-python.svg)
+# pyCSEP: Collaboratory for the Study of Earthquake Predictability
+![](https://i.postimg.cc/Bb60rVQP/CSEP2-Logo-CMYK.png)
+![Python version](https://gist.githubusercontent.com/wsavran/efce311162c32460336a4f9892218532/raw/1b9c060efd1c6e52eb53f82d4249107417d6a5ec/pycsep_python_badge.svg)
 ![Python application](https://github.com/SCECCode/csep2/workflows/Python%20application/badge.svg)
 [![Build sphinx documentation](https://github.com/SCECCode/csep2/workflows/Build%20sphinx%20documentation/badge.svg)](https://cseptesting.org)
 [![codecov](https://codecov.io/gh/SCECcode/pycsep/branch/master/graph/badge.svg?token=HTMKM29MAU)](https://codecov.io/gh/SCECcode/pycsep)
 
-The PyCSEP tools help earthquake forecast model developers evaluate their forecasts with the goal of understanding
+# Description:
+The pyCSEP Toolkit helps earthquake forecast model developers evaluate their forecasts with the goal of understanding
 earthquake predictability.
 
-PyCSEP should:
+pyCSEP should:
 1. Help modelers become familiar with formats, procedures, and evaluations used in CSEP Testing Centers.
 2. Provide vetted software for model developers to use in their research.
-3. Provide quantative and visual tools to assess earthquake forecast quality.
+3. Provide quantitative and visual tools to assess earthquake forecast quality.
 4. Promote open-science ideas by ensuring transparency and availability of scientific code and results.
 5. Curate benchmark models and data sets for modelers to conduct retrospective experiments of their forecasts.
 
-## Using Conda
+# Table of Contents:
+1. [Software Documentation](https://docs.cseptesting.org)
+2. [Installation](#installation)
+3. [Usage](#usage)
+4. [Contributing](#contributing)
+5. [Change Log](https://github.com/SCECcode/pycsep/blob/master/CHANGELOG.txt)
+6. [Credits](#credits)
+7. [License](#license)
 
-The easiest way to install PyCSEP is using `conda`. It can also be installed using `pip` or built from source. 
-If you plan on contributing to this package, visit the 
-[contribution guidelines](https://github.com/SCECcode/pycsep/blob/master/CONTRIBUTING.md) for 
-installation instructions. We recommend creating a `conda` environment for this installation, but this is not a required step.
+# Installation:
+pyCSEP can be installed in several ways. It can be installed using conda or pip package managers or from the 
+source code found in the pyCSEP github repo. Researchers interested in contributing to pyCSEP development should 
+install pyCSEP from source code. pyCSEP depends on the following software packages. 
+These which may be installed automatically, or manually, depending on the installation method used.
+* Python 3.7 or later (https://python.org)
+* NumPy 1.21.3 or later (https://numpy.org)
+* SciPy 1.7.1 or later (https://scipy.org)
+* pandas 1.3.4 or later (https://pandas.pydata.org)
+* cartopy 0.20.0 or later (https://scitools.org.uk/cartopy/docs/latest)
+* GEOS 3.7.2 or later (https://trac.osgeo.org/geos/)
+* PROJ 8.0.0 or later (https://proj.org/)
 
-    conda create -n pycsep
-    conda activate pycsep
-    
-Install `pycsep`
[…]
