### 2021-11-03 4b2a45a
pre-release commit for v0.5.0
- add _version.py
- update changelog to markdown for github formatting
- update codemeta.json with version information
- add sphinx to requirements in setup.py
- add instructions to documentation about issuing releases
