### 2021-11-04 90cc2ec
Modified plot_catalog and plot_spatial_dataset to pass set_global argument to plot_basemap. Does not define an extent in case that set_global is True. Also if set_global, region_border is modified to False to avoid global forecast edges when central_longitude is passed in the projection.
Modified plot_customizations example to pass -179 as central_longitude. Values of 180 or -180 are not correctly handled by cartopy and pcolor functions fails to project appropiately.
