+++ .github/workflows/build-sphinx-no-deploy.yml
@@ -18,12 +18,13 @@ jobs:
           sudo apt-get install libgeos-dev
           sudo apt-get install python3-sphinx
           python -m pip install --upgrade pip
-          pip install numpy
+          pip install setuptools==58.4.0
           pip install wheel
+          pip install numpy
           pip install pytest pytest-cov
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
           pip install pillow
-      - name: Install PyCSEP
+      - name: Install pyCSEP
         run: |
           python setup.py install
       - name: Build documentation
