+++ csep/utils/plots.py
@@ -17,7 +17,7 @@ from cartopy.io import img_tiles
 from csep.utils.constants import SECONDS_PER_DAY, CSEP_MW_BINS
 from csep.utils.calc import bin1d_vec
 from csep.utils.time_utils import datetime_to_utc_epoch
-from csep.core.regions import geographical_area_from_bounds
+from csep.core import regions
 
 
 """
@@ -1641,32 +1641,32 @@ def plot_ROC_Curves(forecasts, catalog, axes=None, name=None, savepdf=False, sav
         N_forecast = 1             # the input forecasts is a single csep.forecast
         forecasts = [forecasts]
     
-    cats_norm_sorted   = np.zeros((N_forecast,len(rate)), dtype=np.float64) # initialze 
+    cats_norm_sorted   = numpy.zeros((N_forecast,len(forecasts[0].spatial_counts())), dtype=numpy.float64) # initialze 
     
     for j,forecast in enumerate(forecasts):
         ### GET area for each geological bin (cell)
         bin_lat = forecast.get_latitudes()                   # bin location in forecast
         bin_lon = forecast.get_longitudes()                  # bin location in forecast
         rate    = forecast.spatial_counts()                  # [eq per cell per duration] in forecast
-        lats = np.unique(bin_lat) 
-        lons = np.unique(bin_lon)
-        min_lat = np.min(lats); max_lat = np.max(lats) # get min/max of grids' lat
-        min_lon = np.min(lons); max_lon = np.max(lons) # get min/max of grids' lon
+        lats = numpy.unique(bin_lat) 
+        lons = numpy.unique(bin_lon)
+        min_lat = numpy.min(lats); max_lat = numpy.max(lats) # get min/max of grids' lat
+        min_lon = numpy.min(lons); max_lon = numpy.max(lons) # get min/max of grids' lon
         d_lat = lats[1] - lats[0]                      # get grid interval [d_lat]
         d_lon = lons[1] - lons[0]                      # get grid interval [d_lon]
 
-        area_km2 = np.zeros(bin_lon.shape, dtype=np.float64) # Initialze
+        area_km2 = numpy.zeros(bin_lon.shape, dtype=numpy.float64) # Initialze
         for i, bot_lat in enumerate(bin_lat):
             bot_lon = bin_lon[i]
             top_lat = bot_lat + d_lat
             top_lon = bot_lon + d_lon
-            area_km2[i] = geographical_area_from_bounds(bot_lon,bot_lat,top_lon,top_lat)
+            area_km2[i] = regions.geographical_area_from_bounds(bot_lon,bot_lat,top_lon,top_lat)
 
         # Get normalized & sorted [prob] (probability) and [area]
-        I = np.argsort(rate)             # get index of ascending sort
-        I = np.flip(I)                   # get index of descending sort
-        area_norm_sorted = np.cumsum(area_km2[I])/ np.sum(area_km2)
-        prob_norm_sorted = np.cumsum(rate[I])    / np.sum(rate)
+        I = numpy.argsort(rate)             # get index of ascending sort
+        I = numpy.flip(I)                   # get index of descending sort
+        area_norm_sorted = numpy.cumsum(area_km2[I])/ numpy.sum(area_km2)
+        prob_norm_sorted = numpy.cumsum(rate[I])    / numpy.sum(rate)
     
         ### GET information score contribution from each testing catalog
         evt_lat = catalog.get_latitudes()    # event location in the testing catalog
[…]
