### 2021-11-04 33b4602
modify documentations (tutorial)
