### 2021-11-06 d3cf2e3
updated catalog forecast evaluation plots with better plot-args added plotting functions to api-reference in docs
