+++ csep/utils/plots.py
@@ -389,7 +389,7 @@ def plot_histogram(simulated, observation, bins='fd', percentile=None,
         pyplot.show()
     return ax
 
-def plot_ecdf(x, ecdf, axes=None, xv=None, show=False, plot_args = None):
+def plot_ecdf(x, ecdf, axes=None, xv=None, show=False, plot_args=None):
     """ Plots empirical cumulative distribution function.  """
     plot_args = plot_args or {}
     # get values from plotting args
@@ -972,32 +972,56 @@ def plot_number_test(evaluation_result, axes=None, show=True, plot_args=None):
     Takes result from evaluation and generates a specific histogram plot to show the results of the statistical evaluation
     for the n-test.
 
-
     Args:
         evaluation_result: object-like var that implements the interface of the above EvaluationResult
+        axes (matplotlib.Axes): axes object used to chain this plot
+        show (bool): if true, call pyplot.show()
+        plot_args(dict): optional argument containing a dictionary of plotting arguments, with keys as strings and items as described below
+
+    Optional plotting arguments:
+        * figsize: (:class:`list`/:class:`tuple`) - default: [6.4, 4.8]
+        * title: (:class:`str`) - default: name of the first evaluation result type
+        * title_fontsize: (:class:`float`) Fontsize of the plot title - default: 10
+        * xlabel: (:class:`str`) - default: 'X'
+        * xlabel_fontsize: (:class:`float`) - default: 10
+        * xticks_fontsize: (:class:`float`) - default: 10
+        * ylabel_fontsize: (:class:`float`) - default: 10
+        * text_fontsize: (:class:`float`) - default: 14
+        * tight_layout: (:class:`bool`) Set matplotlib.figure.tight_layout to remove excess blank space in the plot - default: True
+        * percentile (:class:`float`) Critial region to shade on histogram - default: 95
+        * bins: (:class:`str`) - Set binning type. see matplotlib.hist for more info - default: 'auto'
+        * xy: (:class:`list`/:class:`tuple`) - default: (0.55, 0.3)
 
     Returns:
         ax (matplotlib.axes.Axes): can be used to modify the figure
 
     """
-    plot_args = plot_args or {}
-    # handle plotting
+
+    # chain plotting axes if requested
     if axes:
         chained = True
     else:
         chained = False
-    # supply fixed arguments to plots
-    # might want to add other defaults here
-    filename = plot_args.get('filename', None)
-    xlabel = plot_args.get('xlabel', 'Event count of catalog')
+
+    # default plotting arguments
+    plot_args = plot_args or {}
+    title = plot_args.get('title', 'Number Test')
+    title_fontsize = plot_args.get('title_fontsize', None)
+    xlabel = plot_args.get('xlabel', 'Event count of catalogs')
+    xlabel_fontsize = plot_args.get('xlabel_fontsize', None)
     ylabel = plot_args.get('ylabel', 'Number of catalogs')
+    ylabel_fontsize = plot_args.get('ylabel_fontsize', None)
+    text_fontsize = plot_args.get('text_fontsize', 14)
+    tight_layout = plot_args.get('tight_layout', True)
+    percentile = plot_args.get('percentile', 95)
+    filename = plot_args.get('filename', None)
+    bins = plot_args.get('bins', 'auto')
     xy = plot_args.get('xy', (0.5, 0.3))
 
+    # set default plotting arguments
     fixed_plot_args = {'obs_label': evaluation_result.obs_name,
                        'sim_label': evaluation_result.sim_name}
     plot_args.update(fixed_plot_args)
-    bins = plot_args.get('mag_bins', 'auto')
-    percentile = plot_args.get('percentile', 95)
[…]
