### 2021-11-08 3b33e03
added note about diff results from catalog filtering example included figure thumbnail in documents for catalog filtering
