### 2021-11-08 58ab70e
moved files relative to conf.py changed PyCSEP -> pyCSEP
