### 2021-11-09 65a838d
pre-release commit for v0.5.0 (#149)
* add _version.py
* update changelog to markdown for github formatting
* update codemeta.json with version information
* add sphinx to requirements in setup.py
* add instructions to documentation about issuing releases
 added downloads badges to readme
* added note about diff results from catalog filtering example
included figure thumbnail in documents for catalog filtering
* moved files relative to conf.py
* changed PyCSEP -> pyCSEP
