+++ csep/_version.py
@@ -1,0 +1,2 @@
+__version__ = "0.5.0"
+
