+++ requirements.txt
@@ -9,7 +9,7 @@ python-dateutil
 pytest
 vcrpy
 pytest-cov
-sphinx
+sphinx == 4.2.0
 sphinx-gallery
 sphinx-rtd-theme
 pillow
