+++ requirements.txt
@@ -9,7 +9,7 @@ python-dateutil
 pytest
 vcrpy
 pytest-cov
-sphinx == 4.2.0
-sphinx-gallery
+sphinx
+git+git://github.com/sphinx-gallery/sphinx-gallery@master#egg=sphinx-gallery
 sphinx-rtd-theme
 pillow
