### 2021-11-13 78a86ea
added option to skip drawing colorbar on spatial dataset
