### 2021-11-15 cde2288
updated catalog forecast evaluation plots with better plot-args (#154)
* updated catalog forecast evaluation plots with better plot-args
* added plotting functions to api-reference in docs
* added option to skip drawing colorbar on spatial plot
