+++ csep/utils/plots.py
@@ -389,7 +389,7 @@ def plot_histogram(simulated, observation, bins='fd', percentile=None,
         pyplot.show()
     return ax
 
-def plot_ecdf(x, ecdf, axes=None, xv=None, show=False, plot_args = None):
+def plot_ecdf(x, ecdf, axes=None, xv=None, show=False, plot_args=None):
     """ Plots empirical cumulative distribution function.  """
     plot_args = plot_args or {}
     # get values from plotting args
@@ -883,6 +883,7 @@ def plot_spatial_dataset(gridded, region, ax=None, show=False, extent=None, set_
     linecolor = plot_args.get('linecolor', 'black')
     region_border = plot_args.get('region_border', True)
     # color bar properties
+    include_cbar = plot_args.get('include_cbar', True)
     cmap = plot_args.get('cmap', None)
     clim = plot_args.get('clim', None)
     clabel = plot_args.get('clabel', None)
@@ -937,11 +938,12 @@ def plot_spatial_dataset(gridded, region, ax=None, show=False, extent=None, set_
     # Colorbar options
     # create an axes on the right side of ax. The width of cax will be 5%
     # of ax and the padding between cax and ax will be fixed at 0.05 inch.
-    cax = fig.add_axes([ax.get_position().x1 + 0.01, ax.get_position().y0, 0.025, ax.get_position().height],
-                       label='Colorbar')
-    cbar = fig.colorbar(im, ax=ax, cax=cax)
-    cbar.set_label(clabel, fontsize=clabel_fontsize)
-    cbar.ax.tick_params(labelsize=cticks_fontsize)
+    if include_cbar:
+        cax = fig.add_axes([ax.get_position().x1 + 0.01, ax.get_position().y0, 0.025, ax.get_position().height],
+                           label='Colorbar')
+        cbar = fig.colorbar(im, ax=ax, cax=cax)
+        cbar.set_label(clabel, fontsize=clabel_fontsize)
+        cbar.ax.tick_params(labelsize=cticks_fontsize)
 
     # Gridline options
     if grid:
@@ -972,32 +974,56 @@ def plot_number_test(evaluation_result, axes=None, show=True, plot_args=None):
     Takes result from evaluation and generates a specific histogram plot to show the results of the statistical evaluation
     for the n-test.
 
-
     Args:
         evaluation_result: object-like var that implements the interface of the above EvaluationResult
+        axes (matplotlib.Axes): axes object used to chain this plot
+        show (bool): if true, call pyplot.show()
+        plot_args(dict): optional argument containing a dictionary of plotting arguments, with keys as strings and items as described below
+
+    Optional plotting arguments:
+        * figsize: (:class:`list`/:class:`tuple`) - default: [6.4, 4.8]
+        * title: (:class:`str`) - default: name of the first evaluation result type
+        * title_fontsize: (:class:`float`) Fontsize of the plot title - default: 10
+        * xlabel: (:class:`str`) - default: 'X'
+        * xlabel_fontsize: (:class:`float`) - default: 10
+        * xticks_fontsize: (:class:`float`) - default: 10
+        * ylabel_fontsize: (:class:`float`) - default: 10
+        * text_fontsize: (:class:`float`) - default: 14
+        * tight_layout: (:class:`bool`) Set matplotlib.figure.tight_layout to remove excess blank space in the plot - default: True
+        * percentile (:class:`float`) Critial region to shade on histogram - default: 95
+        * bins: (:class:`str`) - Set binning type. see matplotlib.hist for more info - default: 'auto'
+        * xy: (:class:`list`/:class:`tuple`) - default: (0.55, 0.3)
 
     Returns:
         ax (matplotlib.axes.Axes): can be used to modify the figure
 
     """
[…]
