### 2021-11-15 a31a209
- added function in region to compute cell area - increase timeout for libcomat to 120 seconds - updated docstring for plot_ROC with more information - added plot_args to plot_ROC for customization - modifed input arguments to only accept single forecast per call - simplified code using new function in region class - changed default labels for curves - added support for catalog forecasts - included example of ROC curve for catalog forecast evaluation - simplified I1 score using new function in region class - updated style of code to be consistent with rest of package
