+++ examples/tutorials/gridded_forecast_evaluation.py
@@ -27,6 +27,7 @@ Overview:
 import csep
 from csep.core import poisson_evaluations as poisson
 from csep.utils import datasets, time_utils, plots
+from csep.utils.stats import get_Kagan_I1_score
 
 ####################################################################################################################################
 # Define forecast properties
@@ -59,8 +60,8 @@ forecast = csep.load_gridded_forecast(datasets.helmstetter_aftershock_fname,
 # to filter the catalog in both time and magnitude. See the catalog filtering example, for more information on how to
 # filter the catalog in space and time manually.
 
-catalog = csep.query_comcat(forecast.start_time, forecast.end_time,
-                            min_magnitude=forecast.min_magnitude)
+print("Querying comcat catalog")
+catalog = csep.query_comcat(forecast.start_time, forecast.end_time, min_magnitude=forecast.min_magnitude)
 print(catalog)
 
 ####################################################################################################################################
@@ -107,8 +108,7 @@ ax = plots.plot_poisson_consistency_test(spatial_test_result,
 #
 # We can also get the Kagan's I_1 score (see Kagan, Yan Y. [2009] Testing long-term earthquake forecasts: likelihood methods and error diagrams, Geophys. J. Int., v. 177, pages 532-542.
 
-from csep.utils.stats import get_Kagan_I1_score
-I_1 = get_Kagan_I1_score(forecast,comcat_catalog)
+I_1 = get_Kagan_I1_score(forecast, catalog)
 print("I_1 score is: ", I_1)
 
 ####################################################################################################################################
@@ -118,5 +118,4 @@ print("I_1 score is: ", I_1)
 # We can also plot the Receiver operating characteristic (ROC) Curves based on forecast and testing-catalog.
 # In the figure below, False Positive Rate is the normalized cumulative forecast rate, after sorting cells in decreasing order of rate. The "True Positive Rate" is the normalized cumulative area. The dashed line is the ROC curver of a uniform forecast, meaning the likelihood for an earthquake to occur at any position is the same. The further the ROC curve of a forecast is to the uniform forecast, the specific the forecast is. When comparing the forecast ROC curve against an catalog, one can evaluate if the forecast is more or less specific (or smooth) at different level or seismic rate.
 
-from csep.utils.plots import plot_ROC_Curves
-plot_ROC_Curves(forecast, comcat_catalog)
+plots.plot_ROC(forecast, catalog)
