+++ csep/utils/plots.py
@@ -1728,97 +1728,144 @@ def plot_calibration_test(evaluation_result, axes=None, plot_args=None, show=Fal
 
     return ax
 
-def plot_ROC_Curves(forecasts, catalog, axes=None, name=None, savepdf=False, savepng=False):
+def plot_ROC(forecast, catalog, axes=None, plot_uniform=True, savepdf=True, savepng=True, show=True,
+                    plot_args=None):
     """
-    Plot Receiver operating characteristic (ROC) Curves based on forecast and testing-catalog
+    Plot Receiver operating characteristic (ROC) Curves based on forecast and test catalog.
 
-    Written by Han Bao, UCLA, March 2021. Modified June 2021.
+    The ROC is computed following this procedure:
+        (1) Obtain spatial rates from GriddedForecast
+        (2) Rank the rates in descending order (highest rates first).
+        (3) Sort forecasted rates by ordering found in (2), and normalize rates so the cumulative sum equals unity.
+        (4) Obtain binned spatial rates from observed catalog
+        (5) Sort gridded observed rates by ordering found in (2), and normalize so the cumulative sum equals unity.
+        (6) Compute spatial bin areas, sort by ordering found in (2), and normalize so the cumulative sum equals unity.
+        (7) Plot ordered and cumulated rates (forecast and catalog) against ordered and cumulated bin areas.
 
     Note that: 
-        (1) the testing catalog and forecast should have exactly the same time-window (duration)
-        
+        (1) The testing catalog and forecast should have exactly the same time-window (duration)
+        (2) Forecasts should be defined over the same region
+        (3) If calling this function multiple times, update the color in plot_args
+
     Args:
-        forecasts:  csep.forecast or a list of csep.forecast (one catalog to test against different forecasts)
-        catalog:    csep.catalog (:class:`CSEPCatalog`)
-        axes:       Previously defined ax object (:class:`matplotlib.pyplot.ax`)
-        name:       Title of the figure (:class: `string`)
-        savepdf:    output option of pdf file
-        savepng:    output option of png file
-        
+        forecast (:class: `csep.forecast.GriddedForecast`):
+        catalog (:class:`AbstractBaseCatalog`): evaluation catalog
+        axes (:class:`matplotlib.pyplot.ax`): Previously defined ax object
+        savepdf (str):    output option of pdf file
+        savepng (str):    output option of png file
+        plot_uniform (bool): if true, include uniform forecast on plot
+
+        Optional plotting arguments:
+            * figsize: (:class:`list`/:class:`tuple`) - default: [9, 8]
+            * forecast_linecolor: (:class:`str`) - default: 'black'
+            * title_fontsize: (:class:`float`) Fontsize of the plot title - default: 18
+            * forecast_linecolor: (:class:`str`) - default: 'black'
+            * forecast_linestyle: (:class:`str`) - default: '-'
+            * observed_linecolor: (:class:`str`) - default: 'blue'
+            * observed_linestyle: (:class:`str`) - default: '-'
+            * forecast_label: (:class:`str`) - default: Observed (Forecast)
+            * legend_fontsize: (:class:`float`) Fontsize of the plot title - default: 16
+            * legend_loc: (:class:`str`) - default: 'upper left'
+            * title_fontsize: (:class:`float`) Fontsize of the plot title - default: 18
+            * label_fontsize: (:class:`float`) Fontsize of the plot title - default: 14
[…]
