### 2021-11-15 5ea651d
version bump for v0.5.1 release (#161)
