### 2021-11-16 29c88fa
Add utility functions: (1) plotting ROC curve and (2) function to calculate Kagan_I_1_score (#98)
* Add utility function of plotting ROC curve and function to calculate Kagan_I_1_score

- bug fix
- modify documentations (tutorial)
- added function in region to compute cell area
- increase timeout for libcomat to 120 seconds
- updated docstring for plot_ROC with more information
- added plot_args to plot_ROC for customization
- modifed input arguments to only accept single forecast per call
- simplified code using new function in region class
- changed default labels for curves
- added support 
[…]
