### 2021-11-16 feedd2d
Update CREDITS.md
