### 2021-11-16 99b10a1
clean quadtree_forecast_reader
