### 2021-11-16 6faf15a
Update test_plots.py
