### 2021-11-17 SCECcode/pycsep#163 @mbarall
JOSS Review Thread
1)  I am one of the reviewers assigned by the Journal of Open Source Software. I will use this issue to discuss my experiences with pyCSEP.

I have been aware of CSEP for a long time, but I have never used pyCSEP, and I have never used Python, so I can provide something of a newcomer's perspective.
