### 2021-11-17 SCECcode/pycsep#163 @mbarall
The documentation website is very good. But an essential part of the documentation is paywalled.

The statistical tests used by pyCSEP are defined in six journal articles listed on the references page. It is not possible to understand pyCSEP or use it effectively without knowing what the statistical tests are. All six papers are behind paywalls, and purchasing them individually would cost over $200. (I acknowledge that one of the papers is freely available on Zenodo, and that many likely users will have access to the articles through their institution's library.)

I recommend that you post a freely-available document on your website that defines the statistical tests implemented in pyCSEP, including the relevant mathematical formulas.

(I note that such a document is mentioned in the future development roadmap.)
