### 2021-11-17 SCECcode/pycsep#163 @mbarall
Installation went smoothly.

I created a virtual machine in the cloud, and installed remote desktop software. Then I followed the instructions for installing pyCSEP from source. I installed Miniconda, and then ran the commands shown on the installation page. Everything went smoothly, and all the dependencies were installed automatically.

Good job on the installation.
