### 2021-11-17 SCECcode/pycsep#163 @mbarall
My first attempt to use pyCSEP was to run the example in the README file to download and plot an earthquake catalog from Comcat. I expected it to produce a plot, but no plot appeared. As a first-time Python user, I did not know why there was no plot, if there was something wrong, or what to do about it. I later tried the tutorial to plot a gridded forecast, and again no plot appeared.

Eventually, with the help of some Internet searches, I figured out how to display plots. This might have been obvious to an experienced user, but for me as a first-time user it was frustrating.

I recommend that you amend the tutorials and README to include commands that actually display the plots on-screen. This would provide a better experience for first-time users.
