### 2021-11-17 SCECcode/pycsep#163 @mbarall
I have now run all the tutorials. They all work as expected, except that in most cases they don't include the necessary commands to make plots appear on the screen. They also seem like a good introduction to the capabilities of pyCSEP.
