### 2021-11-17 SCECcode/pycsep#163 @mbarall
The JOSS paper is concise and well-written. Good job.

I recommend that in the summary section, which is intended for a non-specialist audience, you delete the sentence "pyCSEP can evaluate earthquake forecasts expressed as expected rates in space-magnitude bins, and simulation-based forecasts that produce thousands of synthetic seismicity catalogs." That is technical information which belongs in the pyCSEP overview section, and indeed, the overview section already contains that information.
