+++ docs/concepts/regions.rst
@@ -26,16 +26,17 @@ objects provided by PyCSEP. This module contains tools for working with gridded 
 First, we will describe how the spatial regions are handled. Followed by magnitude regions, and how these two aspects
 interact with one another.
 
-**************
-Region objects
-**************
+.. **************
+.. Region objects
+.. **************
 
 Currently, PyCSEP provides two different kinds of spatial gridding approaches to handle binning catalogs and defining regions 
-for earthquake forecasting evaluations, i.e. :class:`CartesianGrid2D<csep.core.regions.CartesianGrid2D>` and :class:`QuadtreeGrid2D<csep.core.regions.QuadtreeGrid2D>`.
+for earthquake forecasting evaluations, i.e. :class:`CartesianGrid2D<csep.core.regions.CartesianGrid2D>` and :class:`CartesianGrid2D<csep.core.regions.QuadtreeGrid2D>`.
 The fruther details about spatial grids are given below.
 
+******************
 2D Cartesian grids
-##################
+******************
 
 This section contains information about using 2D cartesian grids.
 
@@ -97,9 +98,9 @@ mapping defined in (2). The :meth:`get_index_of<csep.core.regions.CartesianGrid2
 of longitudes and latitudes and returns the index of the polygon they are associated with. For instance, this index can
 now be used to access a data value stored in another data structure.
 
-***************
+.. ***************
 Testing Regions
-***************
+########################
 
 CSEP has defined testing regions that can be used for earthquake forecasting experiments. The following functions in the
 :mod:`csep.core.regions` module returns a :class:`CartesianGrid2D<csep.core.regions.CartesianGrid2D>` consistent with
@@ -111,9 +112,9 @@ these regions.
     italy_csep_region
     global_region
 
-****************
+.. ****************
 Region Utilities
-****************
+########################
 
 PyCSEP also provides some utilities that can facilitate working with regions. As we expand this module, we will include
 functions to accommodate different use-cases.
@@ -126,22 +127,29 @@ functions to accommodate different use-cases.
     increase_grid_resolution
     masked_region
     generate_aftershock_region
+
 	
+**************
+Quadtree grid
+**************
 
-Quadtree based grid
-##################
-
-The other gridding approach PyCSEP supports is quadtree-based spatial gridding approach. 
-The quadtree is a hierarchical tiling strategy for storing and indexing geospatial data. In start whole globe is divided into 4 tiles,  identified as '0', '1', '2', '3'.  
+PyCSEP supports is quadtree-based spatial gridding approach. 
+The quadtree is a hierarchical tiling strategy for storing and indexing geospatial data. In start the global testing region is divided into 4 tiles,  identified as '0', '1', '2', '3'.  
 Then each tile can be divided into four children tiles, until a final desired grid is acquired.
-Each tile is identified by a unique identifier called quadkey, which are '0', '1', '2' or '3'. When a tile is divided further, the quadkey is also modified by appending the new number with the previous number. 
-The number of times a tile is divided is reffered as the zoom-level. And the length of quadkey denotes the number of times a tile has been divided. 
-If a single-resolution grid is acquired at zoom-level (L) or 5, it gives 1024 spatial cells in whole globe. Similary at L=11, the number of cells acquired in grid are 4.2 million approx. 
-The quadtree-based gridding approach that is also made capable of providing a multi-resolution spatial grid in addition to single-resolution grid.
-In a multi-resolution grid, the grid resolution is determined by a certain criterion (or multiple criteria), e.g. maximum number of earthquakes allowed per cell (Nmax). 
-This means that only those cells (tiles) are divided further into sub-cells that contain more earthquakes than Nmax. 
[…]
