### 2021-11-17 bb14653
Quadtree example for documentation
