### 2021-11-17 SCECcode/pycsep#163 @kbarnhart
@mbarall thanks for these comments. I'm linking this comment to the JOSS review issue: openjournals/joss-reviews/issues/3658

I don't know whether JOSS has a policy regarding supporting information behind paywalls or not (as you bring up above). I'll reach out to the editorial board and see if there is any specific guidance on that point I can provide.
