### 2021-11-18 SCECcode/pycsep#163 @mbarall
The API documentation looks good. It is a comprehensive description of the classes and functions available in pyCSEP. Although many of the descriptions are quite short, there is enough information so that someone who wants to use pyCSEP in their own project can do so.
