### 2021-11-18 SCECcode/pycsep#163 @mbarall
I installed pytest and then executed the run_tests script. It performed a series of 127 tests of the code, and all the tests passed. Good work.
