### 2021-11-18 a1e673a
minor cleanings-Adjustments to tests
