### 2021-11-19 5155e2a
Docs example - Combine both grids
