+++ examples/tutorials/quadtree_gridded_forecast_evaluation.py
@@ -1,19 +1,29 @@
 """
 
-.. _quadtree_multi-res_grid-forecast-evaluation:
+.. _quadtree_gridded-forecast-evaluation:
 
-Quadtree Multi-resolution Grid-based Forecast Evaluation
+Quadtree Grid-based Forecast Evaluation
 =======================================
 
-This example demonstrates how to create a multi-resolution grid based on earthquake catalog. Then use that grid to create and evaluate a time-independent forecast. Grid-based
+This example demonstrates how to create a quadtree based single resolution-grid and multi-resolution grid.
+Multi-resolution grid is created using earthquake catalog, in which seismic density determines the size of a grid cell. 
+In creating a multi-resolution grid we select a threshold (Nmax) as a maximum number of earthquake in each cell.
+In single-resolution grid, we simply select a zoom-level (L) that determines a single resolution grid.
+The number of cells in single-resolution grid are equal tou 4**L. The zoom-level L=11 leads to 4.2 million cells, nearest to 0.1x0.1 grid.
+
+We use these grids to create and evaluate a time-independent forecast. Grid-based
 forecasts assume the variability of the forecasts is Poissonian. Therefore, Poisson-based evaluations
 should be used to evaluate grid-based forecasts.
 
 Overview:
-    1. Define spatial grid
-    2. Load forecast
+    1. Define spatial grids
+        - Multi-resolution grid
+        - Single-resolution grid
+    2. Load forecasts
+    	- Multi-resolution forecast
+    	- Single-resolution forecast
     3. Load evaluation catalog
-    4. Apply Poissonian evaluations for grid-based forecasts
+    4. Apply Poissonian evaluations for both grid-based forecasts
     5. Visualize evaluation results
 """
 
@@ -33,8 +43,8 @@ from csep.utils.time_utils import decimal_year_to_utc_epoch
 from csep.core.catalogs import CSEPCatalog
 
 ####################################################################################################################################
-# Load Training Catalog
-# --------------------------
+# Load Training Catalog for Multi-resolution grid
+# ----------------------------------------------
 #
 # We define a multi-resolution quadtree using earthquake catalog. We load a training catalog in CSEP and use that catalog to create a multi-resolution grid.
 # Sometimes, we do not the catalog in exact format as requried by PyCSEP. So we can read a catalog using Pandas and convert it 
@@ -65,13 +75,29 @@ print(catalog_train)
 # For simplicity we assume only single magnitude bin, i.e. all the earthquakes greater than and equal to 5.95
 
 mbins = numpy.array([5.95])
-Nmax = 10
-r = QuadtreeGrid2D.from_catalog(catalog_train, Nmax, magnitudes=mbins)
-print('Number of cells in grid :', r.num_nodes)
+Nmax = 25
+r_multi = QuadtreeGrid2D.from_catalog(catalog_train, Nmax, magnitudes=mbins)
+print('Number of cells in Multi-resolution grid :', r_multi.num_nodes)
+
 
 ####################################################################################################################################
-# Load forecast
-# -------------
+# Define Single-resolution Gridded Region
+# ----------------------------------------
+#
+# Here as an example we define a single resolution grid at zoom-level L=6. For this purpose 
+# we call :func:`csep.core.regions.QuadtreeGrid2D_from_single_resolution` to create a single resolution grid.
+
+# For simplicity of example, we assume only single magnitude bin, 
+# i.e. all the earthquakes greater than and equal to 5.95
+
+mbins = numpy.array([5.95])
+r_single = QuadtreeGrid2D.from_single_resolution(6, magnitudes=mbins)
+print('Number of cells in Single-Resolution grid :', r_single.num_nodes)
+
+
+####################################################################################################################################
+# Load forecast of multi-resolution grid
+# ---------------------------------
[…]
