### 2021-11-19 SCECcode/pycsep#163 @kbarnhart
@wsavran, @mbarall A bit of guidance regarding the level of supporting/theory in the documentation based on my conversation with others on the JOSS board.

Project documentation should include enough theory that the user can both run and understand what the software is doing and why. The theory should support user choices (e.g., categorical variable choices like which of a set of options to use and continuous variable choices like what value to give for a parameter).

A point that came up in the editorial board discussion that I want to add as well: while a description of the the theory should be added, it is up to you to determine the extent of rationale, validation, and implications.

E.g., provide description of tests described in Savran et al. (2020) as referenced in the docs [here](https://docs.cseptesting.org/concepts/evaluations.html#catalog-based-forecast-evaluations)
https://pubs.geoscienceworld.org/ssa/bssa/article-abstract/110/4/1799/588154/Pseudoprospective-Evaluation-of-UCERF3-ETAS

I don't suspect this will be too onerous to address. Please let me know if you have questions and/or clarifications.
