### 2021-11-19 #3658 @kbarnhart
@mbarall many thanks for your review!

@wsavran - I would recommend you comment on SCECcode/pycsep/issues/163 once you've addressed the issues outlined there. At that point, @mbarall and I can assess whether you have addressed those concerns.

It is up to you whether you want to address these concerns now, or wait until @nvanderelst has completed his review.

Let me know with any questions, comments, or concerns.
