### 2021-11-19 #3658 @nvanderelst
Hi Katy,
I hoped to have the review completed today, but I have stalled out because I cannot get some of the aspects of the code to work. I'm no python whiz, so I'm probably doing something stupid -- I think Michael got things to work just fine. I will need to work with Bill to figure things out, but this means the review will probably not be finished until after the thanksgiving holiday, as I'm heading into the desert next week.

My apologies!
-Nicholas
