### 2021-11-19 SCECcode/pycsep#163 @wsavran
@mbarall thank you for the review of the code and documentation. i will add technical notes to the documentation that explain the theory behind the tests. i agree that this would stengthen the documentation quite a bit. ill respond here once i have those changes made.
