### 2021-11-19 089e028
updates to documentation and tutorials from joss review * added show option to poisson consistency plots * explicity show figures in tutorials for consistency with terminal users * remove parts from documentation related to ROC and Kagan I1 score (re-add with v0.6.0 release)
