### 2021-11-19 27580bb
updates to documentation and tutorials from joss review * removed extra imports from examples
