### 2021-11-19 111ac34
updates to documentation and tutorials from joss review * added matplotlib show to work until new release
