### 2021-11-19 #3658 @wsavran
@nvanderelst nope, you didn't do anything stupid at all. this one is one me. i merged a new pull request that added new features (ROC plots and Kagan I1 score) including updated documentation. i didn't publish a new release that included those changes, which caused the errors you were getting. i'm fixing that stuff now, and also addressing the trouble with getting the plots to show in the terminal. have fun next week!
