### 2021-11-19 c3853a0
updates to documentation and tutorials from joss review (#165)
* added show option to poisson consistency plots
* explicity show figures in tutorials for consistency with terminal users
* remove parts from documentation related to ROC and Kagan I1 score (re-add with v0.6.0 release)
* removed extra imports from examples
* added matplotlib show to work until new release
