+++ csep/core/poisson_evaluations.py
@@ -234,11 +234,10 @@ def poisson_spatial_likelihood(forecast, catalog):
     
     return poll
 
-
 def binary_spatial_likelihood(forecast, catalog):
     """
     This function computes log-likelihood scores (bills), using a binary likelihood distribution of earthquakes.
-    For this aim, we need an input variable 'forecast' and an variable'catalog'
+    For this aim, we need an input variable 'forecast' and an variable 'catalog'
     
     This function computes the observed log-likehood score obtained by a gridded forecast in each cell, given a
     seismicity catalog. In this case, we assume a binary distribution of earthquakes, so that the likelihood of
@@ -258,13 +257,13 @@ def binary_spatial_likelihood(forecast, catalog):
     X = numpy.zeros(forecast.spatial_counts().shape)
     X[target_idx[0]] = 1
     
-    #First, we estimate the log-likelihood in cells where no events are observed:
+    # First, we estimate the log-likelihood in cells where no events are observed:
     first_term = (1-X) * (-forecast.spatial_counts() * scale)
     
-    #Then, we compute the log-likelihood of observing one or more events given a Poisson distribution, i.e., 1 - Pr(0):
+    # Then, we compute the log-likelihood of observing one or more events given a Poisson distribution, i.e., 1 - Pr(0):
     second_term = X * (numpy.log(1.0 - numpy.exp(-forecast.spatial_counts() * scale)))
     
-    #Finally, we sum both terms to compute log-likelihood score in each spatial cell:
+    # Finally, we sum both terms to compute log-likelihood score in each spatial cell:
     bill = first_term + second_term
     
     return bill
