### 2021-12-08 44808c8
get_xs_ys - Experiment for plotting
