### 2021-12-10 c05bf1f
forecast.plot() for Single-res Quadtree
