### 2022-01-11 856476d
clean up unnecessary comments from code update docstrings for new functions minor text changes to quadtree documentation to fix references and other typos pin numpy version to <= 1.25.0
