+++ csep/core/regions.py
@@ -15,10 +15,6 @@ from csep.utils.scaling_relationships import WellsAndCoppersmith
 
 from csep.models import Polygon
 
-from csep.core.catalogs import CSEPCatalog
-
-
-
 def california_relm_collection_region(dh_scale=1, magnitudes=None, name="relm-california-collection", use_midpoint=True):
     """ Return collection region for California RELM testing region
 
@@ -499,7 +495,6 @@ def _bin_catalog_probability(lons, lats, n_poly, mask, idx_map, binx, biny):
     event_counts[hash_idx] = 1
     return event_counts
 
-
 class CartesianGrid2D:
     """Represents a 2D cartesian gridded region.
 
@@ -786,7 +781,6 @@ def geographical_area_from_bounds(lon1, lat1, lon2, lat2):
     area_km2 = strip_area_steradian * R2 / (360.0 / (lon2 - lon1))
     return area_km2
 
-#--- Quadtree relevant functions and QuadtreeGrid2D class
 def quadtree_grid_bounds(quadk):
     """
     Computes the bottom-left and top-right coordinates corresponding to every quadkey
@@ -852,7 +846,6 @@ def compute_vertices_bounds(bounds, tol=numpy.finfo(float).eps):
         #>>> Top right coords = origin_points[:,2:3]
     """
     return list(map(lambda x: compute_vertex_bounds(x, tol=tol), bounds))
-
 
 def _create_tile(quadk, threshold, zoom, lon, lat, qk, num):
     """
@@ -946,7 +939,6 @@ def _create_tile_fix_len(quadk, zoom, qk):
         #        qk = numpy.append(qk, quadk)
         qk.append(quadk)
 
-
 class QuadtreeGrid2D:
     """
     Respresents a 2D quadtree gridded region. The class provides functionality to generate multi-resolution or single-resolution quadtree grid.
@@ -996,26 +988,29 @@ class QuadtreeGrid2D:
 
     def get_index_of(self, lons, lats): 
         """ Returns the index of lons, lats in self.polygons
+
         Args:
             lons: ndarray-like
             lats: ndarray-like
+
         Returns:
             idx: ndarray-like
         """
-        if isinstance(lons, (list, numpy.ndarray)):  # --If its array or many coords
+        # If its array or many coords
+        if isinstance(lons, (list, numpy.ndarray)):
             idx = []
             for i in range(len(lons)):
                 idx = numpy.append(idx, self._find_location(lons[i], lats[i]))
             idx = idx.astype(int)
-        if isinstance(lons, (int, float)):  # --It its just one Lon/Lon
+            return idx
+        # It its just one Lon/Lon
+        if isinstance(lons, (int, float)):
             idx = self._find_location(lons, lats)
-
-        return idx
+            return idx
+        return None
 
     def _find_location(self, lon, lat):
-        """
-        Takes in single Lon and Lat and finds its Polygon Index.
-        -----Improve this function for End Case scenerios: i.e. Lat = 90 and Lon = -180---
+        """ Takes in single Lon and Lat and finds its Polygon Index.
 
         Returns:
             index number of polyons
@@ -1023,15 +1018,16 @@ class QuadtreeGrid2D:
         loc = numpy.logical_and(numpy.logical_and(lon >= self.bounds[:, 0], lat >= self.bounds[:, 1]),
                                     numpy.logical_and(lon < self.bounds[:, 2], lat < self.bounds[:, 3]))
         if len(numpy.where(loc == True)[0]) > 0:
-            return numpy.where(loc == True)[0][0]  #When find a location. Just a matter for required output
+            return numpy.where(loc == True)[0][0]
         else:
-            return numpy.where(loc == True)[0]  #When find No location. Just a matter for required output
+            return numpy.where(loc == True)[0]
 
[…]
