+++ csep/core/poisson_evaluations.py
@@ -183,10 +183,6 @@ def conditional_likelihood_test(gridded_forecast, observed_catalog, num_simulati
     except CSEPCatalogException:
         observed_catalog.region = gridded_forecast.region
 
-    #if isinstance(gridded_forecast.region, QuadtreeGrid2D): #-----Asim's addition----
-    #    print('Quadtree grid detected.... ')
-    #    gridded_catalog_data = gridded_forecast.region.get_spatial_magnitude_counts(observed_catalog)
-    #else:
     gridded_catalog_data = observed_catalog.spatial_magnitude_counts()
 
     # simply call likelihood test on catalog data and forecast
