### 2022-01-11 378ed35
Quadtree region class - QuadtreeGrid2D (#115)
* QuadtreeGrid2D region added
* Added condition to read QuadtreeGrid forecast
* added spatial_count and spatio_mag_count functions
* poission_evaluations - routed from QuadtreeGrid2D
* modified requirements
* unit tests for quadtree-grid
* removed 'area' from requirements
* modified doc strings
* California quadtreegrid at L=12 added
* california quadtreegrid loading function
* Area function - Taken from Hanbao's code
* setup.py modified-added mercantile
* Area function name/docstring change
* Modified get_cell_area function for quick computation
*
[…]
