+++ csep/core/regions.py
@@ -7,12 +7,13 @@ from xml.etree import ElementTree as ET
 # Third-party imports
 import numpy
 import numpy as np
+import mercantile
 
 # PyCSEP imports
 from csep.utils.calc import bin1d_vec, cleaner_range, first_nonnan, last_nonnan
 from csep.utils.scaling_relationships import WellsAndCoppersmith
-from csep.models import Polygon
 
+from csep.models import Polygon
 
 def california_relm_collection_region(dh_scale=1, magnitudes=None, name="relm-california-collection", use_midpoint=True):
     """ Return collection region for California RELM testing region
@@ -235,7 +236,7 @@ def magnitude_bins(start_magnitude, end_magnitude, dmw):
 
 def create_space_magnitude_region(region, magnitudes):
     """Simple wrapper to create space-magnitude region """
-    if not isinstance(region, CartesianGrid2D):
+    if not (isinstance(region, CartesianGrid2D) or isinstance(region, QuadtreeGrid2D))  :
         raise TypeError("region must be CartesianGrid2D")
     # bind to region class
     if magnitudes is None:
@@ -402,8 +403,8 @@ def compute_vertices(origin_points, dh, tol=numpy.finfo(float).eps):
 
     Notes:
         (x,y) should be accessible like:
-        >>> x_coords = origin_points[:,0]
-        >>> y_coords = origin_points[:,1]
+        #>>> x_coords = origin_points[:,0]
+        #>>> y_coords = origin_points[:,1]
 
     """
     return list(map(lambda x: compute_vertex(x, dh, tol=tol), origin_points))
@@ -493,7 +494,6 @@ def _bin_catalog_probability(lons, lats, n_poly, mask, idx_map, binx, biny):
     # dont accumulate just set to one for probability
     event_counts[hash_idx] = 1
     return event_counts
-
 
 class CartesianGrid2D:
     """Represents a 2D cartesian gridded region.
@@ -780,3 +780,554 @@ def geographical_area_from_bounds(lon1, lat1, lon2, lat2):
                            - 2 * numpy.pi * (1.0e0 - numpy.cos((90.0e0 - lat2) * rad_per_deg))
     area_km2 = strip_area_steradian * R2 / (360.0 / (lon2 - lon1))
     return area_km2
+
+def quadtree_grid_bounds(quadk):
+    """
+    Computes the bottom-left and top-right coordinates corresponding to every quadkey
+
+    Args:
+        qk : Array of Strings
+        Quadkeys.
+
+    Returns:
+        grid_coords : Array of floats
+                    [lon1,lat1,lon2,lat2]
+
+    """
+
+    origin_lat = []
+    origin_lon = []
+    top_right_lon = []
+    top_right_lat = []
+
+    for i in range(len(quadk)):
+        origin_lon.append(mercantile.bounds(mercantile.quadkey_to_tile(quadk[i])).west)
+        origin_lat.append(mercantile.bounds(mercantile.quadkey_to_tile(quadk[i])).south)
+
+        top_right_lon.append(mercantile.bounds(mercantile.quadkey_to_tile(quadk[i])).east)
+        top_right_lat.append(mercantile.bounds(mercantile.quadkey_to_tile(quadk[i])).north)
+
+    grid_origin = numpy.column_stack((numpy.array(origin_lon), numpy.array(origin_lat)))
+    grid_top_right = numpy.column_stack((numpy.array(top_right_lon), numpy.array(top_right_lat)))
+    grid_bounds = numpy.column_stack((grid_origin, grid_top_right))
+
+    return grid_bounds
+
+def compute_vertex_bounds(bound_point, tol=numpy.finfo(float).eps):
+    """
[…]
