+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy
+numpy <= 1.25.0
 scipy
 pandas
 matplotlib
@@ -13,3 +13,5 @@ sphinx
 sphinx-gallery
 sphinx-rtd-theme
 pillow
+mercantile
+
