### 2022-01-12 46132f0
added documentation for roc and i1 score
