+++ .github/workflows/python-app.yml
@@ -7,7 +7,7 @@ on:
   push:
     branches: [ master ]
   pull_request:
-    branches: [ master ]
+    branches: 'v*'
 
 jobs:
   build:
