+++ docs/getting_started/theory.rst
@@ -1,0 +1,1095 @@
+Theory of CSEP Tests
+====================
+
+This page describes the theory of each of the forecast tests
+included in pyCSEP along with working code examples. You will find
+information on the goals of each test, the theory behind the tests, how
+the tests are applied in practice, and how forecasts are ‘scored’ given
+the test results. Also, we include the code required to run in each test
+and a description of how to interpret the test results.
+
+.. code:: ipython3
+
+    import csep
+    from csep.core import (
+        regions,
+        catalog_evaluations,
+        poisson_evaluations as poisson
+    )
+    from csep.utils import (
+        datasets,
+        time_utils,
+        comcat,
+        plots,
+        readers
+    )
+
+    # Filters matplotlib warnings
+    import warnings
+    warnings.filterwarnings('ignore')
+
+Grid-based Forecast Tests
+-------------------------
+
+These tests are designed for grid-based forecasts (e.g., Schorlemmer et
+al., 2007), where expected rates are provided in discrete Poisson
+space-magnitude cells covering the region of interest. The region
+:math:`\boldsymbol{R}` is then the product of the spatial rate
+:math:`\boldsymbol{S}` and the binned magnitude rate
+:math:`\boldsymbol{M}`,
+
+.. math::  \boldsymbol{R} = \boldsymbol{M} \times \boldsymbol{S}.
+
+A forecast :math:`\boldsymbol{\Lambda}` can be fully specified as the
+expected number of events (or rate) in each space-magnitude bin
+(:math:`m_i, s_j`) covering the region :math:`\boldsymbol{R}` and
+therefore can be written as
+
+.. math::  \boldsymbol{\Lambda} = \{ \lambda_{m_i, s_j}| m_i \in \boldsymbol{M}, s_j \in \boldsymbol{S} \},
+
+where :math:`\lambda_{m_i, s_j}` is the expected rate of events in
+magnitude bin :math:`m_i` and spatial bin :math:`s_j`. The observed
+catalogue of events :math:`\boldsymbol{\Omega}` we use to evaluate the
+forecast is similarly discretised into the same space-magnitude bins,
+and can be described as
+
+.. math::  \boldsymbol{\Omega} = \{ \omega_{m_i, s_j}| m_i \in \boldsymbol{M}, s_j \in \boldsymbol{S} \},
+
+where :math:`\omega_{m_i, s_j}` is the observed number of
+events in spatial cell :math:`s_j` and magnitude bin :math:`m_i`. The
+magnitude bins are specified in the forecast: typically these are in 0.1
+increments and this is the case in the examples we use here. These
+examples use the Helmstetter et al (2007) smoothed seismicity forecast
+(including aftershocks), testing over a 5 year period between 2010 and
+2015.
+
+.. code:: ipython3
+
+    # Set up experiment parameters
+    start_date = time_utils.strptime_to_utc_datetime('2010-01-01 00:00:00.0')
+    end_date = time_utils.strptime_to_utc_datetime('2015-01-01 00:00:00.0')
+
+    # Loads from the PyCSEP package
+    helmstetter = csep.load_gridded_forecast(
+        datasets.helmstetter_aftershock_fname,
+        start_date=start_date,
+        end_date=end_date,
+        name='helmstetter_aftershock'
+    )
+
+    # Set up evaluation catalog
+    catalog = csep.query_comcat(helmstetter.start_time, helmstetter.end_time,
+                                min_magnitude=helmstetter.min_magnitude)
+
+    # Filter evaluation catalog
+    catalog = catalog.filter_spatial(helmstetter.region)
+
+    # Add seed for reproducibility in simulations
+    seed = 123456
+
+    # Number of simulations for Poisson consistency tests
+    nsim = 100000
+
+
+.. parsed-literal::
+
[…]
