### 2022-01-13 3c29710
updated uri for figures
