### 2022-01-13 e200cfe
fixes to docs (#172)
* add documentation for testing
* add notebooks folder for cookbook recipes
