### 2022-01-13 a521db5
Update theory.rst
