### 2022-01-13 #3658 @wsavran
@kbarnhart I think I have addressed all of the comments from the reviewers. I'll summarize them here:
1. added a [page](https://docs.cseptesting.org/getting_started/theory.html) on the documentation explaining the mathematical theory behind the evaluations
2. updated the tutorials so they display plots properly when run via the terminal
3. added a new branching model to ensure that documentation changes are not pushed into production that could break the tutorials

Thanks for managing this process and thank you to the reviewers for working through the software package.
