### 2022-01-18 0d67c80
Fixed dh issue by modifying get_bbox()
