### 2022-01-18 SCECcode/pycsep#163 @kbarnhart
@kbarnhart the new docs look great. While looking them over I noticed that there are some heading formatting issues (e.g., CL-test not a heading). Obviously a minor concern relative to the content, but I'd recommend addressing them because they influence the left hand column navigation.
