### 2022-01-18 #3658 @kbarnhart
@wsavran I've looked over your recent changes and it looks to me that you've addressed issues identified by the reviewers. I made one minor comment about documentation in that issue thread.

At this point in the JOSS review process what happens is that @mbarall and @nvanderelst revisit their reviews in light of your changes. If they find that you have addressed their concerns, they close any issues they opened and complete their checklists above. If they find anything that they still think should be addressed, they can point it out. I'll write a quick email to each of them now indicating they should do this.

Once they have revisited and completed their reviews, we enter the final stage of the JOSS review process which involves a light copyedit from me and updating of some metadata. I then pass off handling to one of our managing editors in chief.
