### 2022-01-19 3337bf2
Fixed dh issue by modifying get_bbox() (#175)
* Fixed dh issue by modifying get_bbox()
* deleting extra comments
