### 2022-01-19 0868964
Update theory.rst
