### 2022-01-19 SCECcode/pycsep#163 @wsavran
@kbarnhart i agree. thanks for catching that! i pushed out a fix for the bugged up header.
