### 2022-01-20 SCECcode/pycsep#163 @mbarall
@wsavran Thank you for the new documentation on the theory of CSEP tests.  It's great to have descriptions of the tests all together in one place, and I appreciate that you also included code to invoke each test.

I re-ran all the tutorials, and verified that they all display plots now when run from command line

One quibble: In README.md, the example to download and plot an earthquake catalog still does not display the plot when run from command line, at least not for me.
