### 2022-01-20 SCECcode/pycsep#163 @wsavran
@mbarall Thanks for looking through everything! I just updated the snippet in README.md to show the plot from the command line. Thanks for pointing that out!
