### 2022-01-21 SCECcode/pycsep#163 @mbarall
@wsavran Thank you for updating the README.md snippet.  It works for me now.

Feel free to close this issue at your discretion.
