### 2022-01-21 #3658 @mbarall
@kbarnhart I am satisfied with the changes.  I recommend that you accept pyCSEP for publication.
