### 2022-01-24 #3658 @mjw98
> @kbarnhart I am satisfied with the changes. I recommend that you accept pyCSEP for publication.

Thank you, @mbarall , for the thorough review.
