### 2022-01-24 #3658 @kbarnhart
@wsavran @mjw98 and co-authors:

Please address the following:
- [ ] fix the missing and invalid DOIs listed above
- [ ] consistent formatting of pyCSEP in the manuscript (e.g., line 12 different from line 16)
- [ ] remove the "type" entry in the .bib file (e.g., type = {Journal Article})
- [ ] line 165: I think the i before the colon should be capitalized (enclose in curly braces)

Sorry to miss some of those reference issues on my first run through with the paper.

Once you've addressed those minor issues, please re-run
- `@whedon check references from branch pycsep_joss`
- `@whedon generate pdf from branch pycsep_joss`

and verify that and verify that no missing or invalid DOIs are returned and that the PDF compiles without issue.

At that point, please do the following:

Once all these steps are completed, I will update the paper metadata and recommend that the submission be accepted. At that point whedon will do a dry run of paper publishing and the JOSS managing editor in chief will handle the final steps.

If at any point you have questions for me, please reach out here or at krbarnhart@usgs.gov
