### 2022-01-25 7d624c8
bumped version and added new contributors updated CHANGELOG.md updated CREDITS.md updated conf.py fixed version number of numpy to fix build issues
