+++ docs/conf.py
@@ -52,7 +52,6 @@ extensions = [
     'sphinx.ext.intersphinx',
     'sphinx_gallery.gen_gallery',
     'sphinx.ext.githubpages'
-
 ]
 
 # Add any paths that contain templates here, relative to this directory.
