+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy=1.25.5
+numpy<=1.21.5
 scipy
 pandas
 matplotlib
