### 2022-01-25 8e242b1
updates from joss review
