+++ paper.bib
@@ -4,9 +4,8 @@
    journal = {Seismological Research Letters},
    pages = {1-10},
    url = {http://www.cseptesting.org/documents/files/relm/field.pdf},
-   DOI = {https://doi.org/10.1785/gssrl.78.1.7},
-   year = {2007},
-   type = {Journal Article}
+   DOI = {10.1785/gssrl.78.1.7},
+   year = {2007}
 }
 
 @article{Helmstetter2007a,
@@ -17,8 +16,7 @@
    number = {1},
    pages = {78-86},
    ISSN = {1938-2057},
-   year = {2007},
-   type = {Journal Article}
+   year = {2007}
 }
 
 @article{Schorlemmer2007a,
@@ -28,9 +26,9 @@
    volume = {78},
    number = {1},
    pages = {17-29},
+   DOI = {10.1785/gssrl.78.1.17},
    url = {http://srl.geoscienceworld.org/content/78/1/17.extract},
-   year = {2007},
-   type = {Journal Article}
+   year = {2007}
 }
 
 @article{Schorlemmer2007b,
@@ -43,8 +41,7 @@
    ISSN = {0895-0695},
    DOI = {10.1785/gssrl.78.1.30},
    url = {https://pubs.geoscienceworld.org/srl/article/78/1/30-36/143321},
-   year = {2007},
-   type = {Journal Article}
+   year = {2007}
 }
 
 @article{Zechar2009a,
@@ -56,8 +53,7 @@
    pages = {1836-1847},
    ISSN = {1532-0626},
    DOI = {10.1002/cpe.1519},
-   year = {2009},
-   type = {Journal Article}
+   year = {2009}
 }
 
 @article{Jordan2011a,
@@ -67,8 +63,7 @@
    DOI = {10.4401/ag-5350},
    url = {https://www.annalsofgeophysics.eu/index.php/annals/article/view/5350
 https://www.annalsofgeophysics.eu/index.php/annals/article/download/5350/5371},
-   year = {2011},
-   type = {Journal Article}
+   year = {2011}
 }
 
 @article{Rhoades2011a,
@@ -82,8 +77,7 @@ https://www.annalsofgeophysics.eu/index.php/annals/article/download/5350/5371},
    DOI = {10.2478/s11600-011-0013-5},
    url = {<Go to ISI>://WOS:000290032300005
 https://link.springer.com/content/pdf/10.2478%2Fs11600-011-0013-5.pdf},
-   year = {2011},
-   type = {Journal Article}
+   year = {2011}
 }
 
 @article{Werner2011a,
@@ -96,13 +90,12 @@ https://link.springer.com/content/pdf/10.2478%2Fs11600-011-0013-5.pdf},
    ISSN = {0037-1106},
    DOI = {10.1785/0120090340},
    url = {https://dx.doi.org/10.1785/0120090340},
-   year = {2011},
[…]
