### 2022-01-25 a382346
v0.5.2 release commit (#178)
* bumped version and added new contributors
updated CHANGELOG.md
updated CREDITS.md
updated conf.py
fixed version number of numpy to fix build issues
