+++ csep/_version.py
@@ -1,2 +1,2 @@
-__version__ = "0.5.1"
+__version__ = "0.5.2"
 
