+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy
+numpy<=1.21.5
 scipy
 pandas
 matplotlib
