### 2022-01-25 28760ce
add apt-get update to build script
