### 2022-01-25 cdd84f0
updated developer notes with release instructions
