### 2022-01-25 #3658 @kbarnhart
@wsavran what's the final version number? v0.5.2, yes?
