### 2022-01-26 #3658 @wsavran
@kbarnhart yes, v0.5.2, is the latest release
