### 2022-01-26 #3658 @kbarnhart
@wsavran I've recommended that this paper be accepted. One of the @openjournals/joss-eics will handle the submission from here. Congratulations on contributing a valuable piece of research software. Many thanks to @nvanderelst and @mbarall for your reviews.
